==> 00_testdocument.Rmd <==
---
title: "00_testdocument"
author: "Laura Botzet, Sarah Forsthoff, Tanja Gerlach"
date: "20 April 2018"
output: html_document
---

## Library

```{r}
library(survival)
library(coxme)
```



simulate data using proposed code
```{r}
########## simulate data ##########
set.seed(10)
N <- 250
dat <- data.frame(ID = factor(1:N), age = rnorm(N, mean = 45, sd = 5), sex = sample(0:1, 
    N, TRUE), basemort = rnorm(N, sd = 3))

interval <- matrix(sample(2:14, N * 3, replace = TRUE), N)
windows <- t(apply(cbind(0, interval), 1, cumsum))
windows <- rbind(windows[, 1:2], windows[, 2:3], windows[, 3:4])

colnames(windows) <- c("time1", "time2")
dat <- cbind(do.call(rbind, rep(list(dat), 3)), windows)
dat <- dat[order(dat$ID), ]
dat$assessment <- rep(1:3, N)
rownames(dat) <- NULL
head(dat)
##   ID  age sex basemort time1 time2 assessment
## 1  1 45.1   1    6.659     0    14          1
## 2  1 45.1   1    6.659    14    16          2
## 3  1 45.1   1    6.659    16    23          3
## 4  2 44.1   1    0.943     0     8          1
## 5  2 44.1   1    0.943     8    20          2
## 6  2 44.1   1    0.943    20    22          3
# simulate survival (mortality) data
transplant <- with(dat, {
    mu <- (0.05 * age) + (0.3 * time2)
    lp <- rnorm(N * 3, mean = mu, sd = 1)
    as.integer(lp > quantile(lp, probs = 0.65))
})
# ensure that transplants do not revert
transplant <- as.integer(ave(transplant, dat$ID, FUN = cumsum) >= 1)

# simulate survival (mortality) data
mortality <- with(dat, {
    mu <- basemort + (0.05 * age) - (2.5 * sex) + (0.3 * time2)
    lp <- rnorm(N * 3, mean = mu, sd = 1)
    as.integer(lp > median(lp))
})

# ensure that once someone dies, he or she stays dead
mortality <- as.integer(ave(mortality, dat$ID, FUN = cumsum) >= 1)

# ensure no one dead at baseline
mortality[dat$assessment == 1] <- 0

# ensure no post mortem change in transplant status
transplant <- unlist(by(data.frame(mortality, transplant), dat$ID, FUN = function(x) {
    i <- cumsum(x$mortality)
    tstat <- x$transplant[i == 1]
    x$transplant[i >= 1] <- tstat
    return(x$transplant)
}))

dat$transplant <- transplant
dat$mortality <- mortality

# print first few rows
head(dat)
```

```{r}
########## basic models ########## simple model
m <- coxph(Surv(time1, time2, mortality) ~ age + sex + transplant, data = dat)
## summary of the model
summary(m)
plot(survfit(m))
```


```{r}
## model with robust SE via clustering
m2 <- coxph(Surv(time1, time2, mortality) ~ age + sex + transplant + cluster(ID), 
    data = dat)
## summary of the model
summary(m2)
plot(survfit(m2))
```


## test data

```{r}
testdata <- read.csv("C:/Users/q/Desktop/Speicherung/Arbeit/3/Narc_Relationship/narq_relationship_transitions/testdata.csv")


########## basic models ########## simple model
m <- coxph(Surv(Single_Start, Beziehung_Start, Beziehung_event) ~ Age + Male + NARQ_Adm, data = testdata)
## summary of the model
summary(m)
plot(survfit(m))

####model with robust SE via clustering
 
m <- coxph(Surv(Single_Start, Beziehung_Start, Beziehung_event) ~ Age + Male + NARQ_Adm + cluster(Session_Code),
           data = testdata)
## summary of the model


summary(m)
plot(m)
```
