==> 01_dataimport.Rmd <==
---
title: "01_datawrangling"
author: "Laura Botzet, Sarah Forsthoff, Tanja Gerlach"
output: html_document
---

## Library
```{r}
library(jsonlite)
library(formr)
library(psych)
library(stringr)
library(dplyr)
```

## Daten importieren
```{r}

setwd("C:/Users/q/Desktop/Speicherung/Arbeit/3/Narc_Relationship/narq_relationship_transitions") # set workingdirectory

pw2015_demographie = as.data.frame(fromJSON("Data/pw2015_demographie.json"))
pw2015_items = as.data.frame(fromJSON("Data/pw2015_items.json"))
pw2015_t2 = as.data.frame(fromJSON("Data/pw2015_t2.json"))
pw2015_t2_frauen = as.data.frame(fromJSON("Data/pw2015_t2_frauen.json"))
pw2015_t2_ende = as.data.frame(fromJSON("Data/pw2015_t2_ende.json"))
pw2015_t3_start = as.data.frame(fromJSON("Data/pw2015_t3_start.json"))
pw2015_t3_singles = as.data.frame(fromJSON("Data/pw2015_t3_singles.json"))
pw2015_t3_relationship_number = as.data.frame(fromJSON("Data/pw2015_t3_relationship_number.json"))
#pw2015_t3_rate_partners = as.data.frame(fromJSON("Data/pw2015_t3_rate_partners.json"))
#Meiner Meinung nach brauche ich keine Variablen aus dem sheet "_rate_partners"
pw2015_t3_partners = as.data.frame(fromJSON("Data/pw2015_t3_partners.json"))
pw2015_t3_mainrelation = as.data.frame(fromJSON("Data/pw2015_t3_mainrelation.json"))
pw2015_t3_ende = as.data.frame(fromJSON("Data/pw2015_t3_ende.json"))
```

## Data Wrangling
Kurzvariable für erste 10 Zeichen des Session-Codes erstellen
```{r}
pw2015_demographie$short_session = str_sub(pw2015_demographie$session,1,10)
pw2015_items$short_session = str_sub(pw2015_items$session,1,10)
pw2015_t2$short_session = str_sub(pw2015_t2$session,1,10)
pw2015_t2_frauen$short_session = str_sub(pw2015_t2_frauen$session,1,10)
pw2015_t2_ende$short_session = str_sub(pw2015_t2_ende$session,1,10)
pw2015_t3_start$short_session = str_sub(pw2015_t3_start$session,1,10)
pw2015_t3_singles$short_session = str_sub(pw2015_t3_singles$session,1,10)
pw2015_t3_relationship_number$short_session = str_sub(pw2015_t3_relationship_number$session,1,10)
#pw2015_t3_rate_partners$short_session = str_sub(pw2015_t3_rate_partners$session,1,10)
pw2015_t3_partners$short_session = str_sub(pw2015_t3_partners$session,1,10)
pw2015_t3_mainrelation$short_session = str_sub(pw2015_t3_mainrelation$session,1,10)
pw2015_t3_ende$short_session = str_sub(pw2015_t3_ende$session,1,10)
```


## Zusammenstellen der benötigten Variablen
Die System- bzw. neu gebildeten Variablen Variablen "session", "short_session" und "ended" macht immer Sinn drin zu lassen. Erstere werden für die Zuordnung gebraucht, letztere erlaubt es Abbrecher herauszunehmen. Falls es in dem Sheet eine Variable gibt die "notes_to_us" oder "comment" heisst auch drin lassen. Die brauchen wir später, um z.B. Fälle, bei denen es technische Probleme oder Verständnisschwierigkeiten gab, herauszufiltern

### t1
```{r}
pw2015_demographie_select = pw2015_demographie %>%
  select(session, short_session, created, ended, sex, age, occupation, education, sex_orientation,
         attracted_to, starts_with("NARQ_"), starts_with("BFIK"))


pw2015_items_select = pw2015_items %>%
  select(session, short_session, created, ended,
         number_relationships, end_relationship_weeks, end_relationship_months, 
         end_relationship_years,
         duration_relationship_years, duration_relationship_months,
         starts_with("global_interest_"),
         starts_with("SOI_R"),
         starts_with("longterm_"),
         notes_to_us,
         email2)
```


### t2
```{r}
pw2015_t2_select = pw2015_t2 %>%
  select(session, short_session, created, ended, 
         starts_with("relationship"),
         starts_with("pastrelationship"))


pw2015_t2_frauen_select = pw2015_t2_frauen %>%
  select(session, short_session, created, ended,
         starts_with("relationship"),
         starts_with("pastrelationship"))


pw2015_t2_ende_select = pw2015_t2_ende %>%
  select(session, short_session, created, ended, notes_to_us_t2)
```



### t3
```{r}
pw2015_t3_start_select = pw2015_t3_start %>%
  select(session, short_session, created, ended, F_self_t3_life_events,F_self_t3_life_event_other)


pw2015_t3_singles_select = pw2015_t3_singles %>%
  select(session, short_session, created, ended, F_self_t3_interest_longterm_single_now)


pw2015_t3_relationship_number_select = pw2015_t3_relationship_number %>%
  select(session, short_session, created, ended, F_t3_relationship_ultimate, F_t3_relationship)


pw2015_t3_partners_select = pw2015_t3_partners %>%
  select(session, short_session, created, ended, F_t3_start, F_t3_lasting, F_t3_end, F_t3_breakup,
         F_t3_kind, F_t3_past_kind, F_t3_closeup_kind, F_t3_name, F_t3_name_new, F_t3_age_new,
         F_t3_sex_new)


pw2015_t3_mainrelation_select = pw2015_t3_mainrelation %>%
  select(session, short_session, created, ended,
         starts_with("F_self_t3_IMS"), F_mate_t3_IMS_alternaavailab,
         starts_with("F_self_t3_transgreact"), 
         F_t3_mainrelation_sex, F_t3_mainrelation_age, 
         F_t3_mainrelation_description)

pw2015_t3_ende_select = pw2015_t3_ende %>%
  select(session, short_session, created, ended, F_t3_honesty_check, F_t3_dishonest_answers, F_t3_notes_to_us)
```


## Abspeichern von Datenfiles
```{r}
setwd("C:/Users/q/Desktop/Speicherung/Arbeit/3/Narc_Relationship/narq_relationship_transitions/ImportedData")

options(encoding = "utf8")

write_json(pw2015_demographie_select, "pw2015_demographie_select.json")
write_json(pw2015_items_select, "pw2015_items_select.json")
write_json(pw2015_t2_select, "pw2015_t2_select.json")

pw2015_t2_frauen_select = pw2015_t2_frauen_select %>%
  mutate(pastrelationship5_f_name = "missing",
         pastrelationship5_name_f_new = "missing",
         pastrelationship5_age_f_new = "missing",
         pastrelationship5_height_f_new = "missing",
         pastrelationship5_relationship_f_new = "missing",
         pastrelationship5_duration_f_new = "missing",
         pastrelationship5_kind_f = "missing",
         pastrelationship5_start_f = "missing",
         pastrelationship5_end_f = "missing",
         pastrelationship6_f = "missing",
         pastrelationshipmore_f = "missing")

write_json(pw2015_t2_frauen_select, "pw2015_t2_frauen_select.json")
write_json(pw2015_t2_ende_select, "pw2015_t2_ende_select.json")
write_json(pw2015_t3_start_select, "pw2015_t3_start_select.json")
write_json(pw2015_t3_singles_select, "pw2015_t3_singles_select.json")
write_json(pw2015_t3_relationship_number_select, "pw2015_t3_relationship_number_select.json")
write_json(pw2015_t3_partners_select, "pw2015_t3_partners_select.json")
write_json(pw2015_t3_mainrelation_select, "pw2015_t3_mainrelation_select.json")
write_json(pw2015_t3_ende_select, "pw2015_t3_ende_select.json")
```
