==> 02_alt_datawrangling.Rmd <==
---
title: "alt_02_datawrangling"
author: "Laura Botzet, Sarah Forsthoff, Tanja Gerlach"
date: "20 April 2018"
output: html_document
---


## Open questions:
- Check To-Do's for Analyses -> 03_analyses
- Get age from t2_partners already assessed at T1 
- Analyses of relationship dissolution after relationship start

@laura:
- Create start_single variable (s. 03_analyses): Can you reuse previous code?
- how to interpret survival summery?

@sarah:
DONE (- t2-alt. with updated datawrangling code)
DONE (- survival analysis for relationship end)
- how to interpret survival summery?
- identical data with both datawrangling scripts?? I used "all.equal". Maybe check results together


DONE
    To do: Transgression survival analysis for relationship end
    - "event"-variable
    - "end"-date for censored data (if relationship (lasting), then created t2/t3)



## Library
```{r}
# library(jsonlite)
# install.packages("devtools")
# devtools::install_github("rubenarslan/formr")
 library(formr)
# library(psych)
# library(stringr)
library(mosaic)
library(tidyr)
library(apaTables)
library(dplyr)
```



## Data Import
```{r}
# setwd("C:/Users/q/Desktop/Speicherung/Arbeit/3/Narc_Relationship/narq_relationship_transitions")
#setwd("/Users/sarahforsthoff/Documents/Bachelor Arbeit/Hiwi/narq_relationship_transitions")
# set workingdirectory

pw2015_demographie =  read.csv("ImportedData/pw2015_demographie_select.csv", sep = ";")
pw2015_items = read.csv("ImportedData/pw2015_items_select.csv", sep = ";")
pw2015_t2 = read.csv("ImportedData/pw2015_t2_select.csv", sep = ";")
pw2015_t2_frauen = read.csv("ImportedData/pw2015_t2_frauen_select.csv", sep = ";")
pw2015_t2_ende = read.csv("ImportedData/pw2015_t2_ende_select.csv", sep = ";")
#pw2015_t3_start = read.csv("ImportedData/pw2015_t3_start_select.csv", sep = ";")
pw2015_t3_singles = read.csv("ImportedData/pw2015_t3_singles_select.csv", sep = ";")
# pw2015_t3_relationship_number = read.csv("ImportedData/pw2015_t3_relationship_number_select.csv", sep = ";")
pw2015_t3_partners = read.csv("ImportedData/pw2015_t3_partners_select.csv", sep = ";")
# pw2015_t3_mainrelation = read.csv("ImportedData/pw2015_t3_mainrelation_select.csv", sep = ";")
pw2015_t3_ende = read.csv("ImportedData/pw2015_t3_ende_select.csv", sep = ";")
```

## Clean Data (Testpeople, Duplicates, Liars, etc.)
### Testperson and missing session
#### t1
```{r}
# Demographie
## Testpeople (all individuals including a "XXX" in their session code, except: aldt4TnU6Yj8G_EP8PuXXXkLfBmn-03L_NRLpKrR0IRNqKlxIWGQA7ED3yvWF-rZ)
test_demographie = pw2015_demographie %>%
  filter((session %contains% "XXX") &
         session != "aldt4TnU6Yj8G_EP8PuXXXkLfBmn-03L_NRLpKrR0IRNqKlxIWGQA7ED3yvWF-rZ")
test_demographie_vector = as.character(test_demographie$session)

# Items
## Testpeople
test_items = pw2015_items %>%
  filter((session %contains% "XXX") &
         session != "aldt4TnU6Yj8G_EP8PuXXXkLfBmn-03L_NRLpKrR0IRNqKlxIWGQA7ED3yvWF-rZ")
test_items_vector = as.character(test_items$session)

## notes to us
comments_pw2015_items = pw2015_items %>%
  filter(notes_to_us %contains% "Cyril Test"|
           notes_to_us %contains% "Tabea Test!"|
           notes_to_us == "Test!" |
           notes_to_us %contains% "Test, test, test." |
           notes_to_us == "Hi Eva und Tabea, schubst mich mal weiter :), damit ich morgen früh gleich an T2 teilnehmen kann. \nIch bin mir ziemlich sicher, dass ich diesmal ausschliesslich heterosexuell ausgefüllt habe - trotzdem musste ich gerade das Frauennetzwerk ausfüllen. Vielleicht prüft ihr das nochmal. \nLG, Tanja" |
           notes_to_us == "Test Eva" |
           notes_to_us == "Test, Eva" |
           notes_to_us == "Test!!! Silvia B." |
           notes_to_us == "selina testet" |
           notes_to_us %contains% "thank the formr monkey\nmany times" |
           notes_to_us == "Test Silvia" |
           notes_to_us == "Test Silvi")

comments_pw2015_items_vector = as.character(comments_pw2015_items$session)
```

"When was your last relationship?": Exclude participants with "funny" answers = last relationshin when younger than 10
```{r}
#merge t1 files including age and time since last relationship
t1_age <- pw2015_demographie %>% select(session, short_session, age)

table(pw2015_items$end_relationship_weeks)
table(pw2015_items$end_relationship_months) 
#No entries exceeding 52 (weeks) or 12 (months), therefore only years will be checked
t1_time_single <- pw2015_items %>% select(session, short_session, end_relationship_years)

t1_time_single_test <- left_join(t1_age, t1_time_single)
t1_time_single_test <- t1_time_single_test %>% filter((age - end_relationship_years) < 10) #three cases detected

vec_time_single_test <- t1_time_single_test$session #vector with session code of the three cases

#exclude from all data
pw2015_demographie =  pw2015_demographie %>%
  filter(!(session %in% vec_time_single_test))

pw2015_items =  pw2015_items %>%
  filter(!(session %in% vec_time_single_test))  

pw2015_t2 =  pw2015_t2 %>%
  filter(!(session %in% vec_time_single_test))
  
pw2015_t2_frauen =  pw2015_t2_frauen %>%
  filter(!(session %in% vec_time_single_test))

pw2015_t2_ende =  pw2015_t2_ende %>%
  filter(!(session %in% vec_time_single_test))

pw2015_t3_singles =  pw2015_t3_singles %>%
  filter(!(session %in% vec_time_single_test))

pw2015_t3_partners =  pw2015_t3_partners %>%
  filter(!(session %in% vec_time_single_test))

pw2015_t3_ende =  pw2015_t3_ende %>%
  filter(!(session %in% vec_time_single_test))

```


#### t2
``` {r}
test_pw2015_t2 = pw2015_t2 %>%
  filter((session %contains% "XXX") &
         session != "aldt4TnU6Yj8G_EP8PuXXXkLfBmn-03L_NRLpKrR0IRNqKlxIWGQA7ED3yvWF-rZ")
test_pw2015_t2_vector = as.character(test_pw2015_t2$session)

test_pw2015_t2_frauen = pw2015_t2_frauen %>%
  filter((session %contains% "XXX") &
         session != "aldt4TnU6Yj8G_EP8PuXXXkLfBmn-03L_NRLpKrR0IRNqKlxIWGQA7ED3yvWF-rZ")
test_pw2015_t2_frauen_vector = as.character(test_pw2015_t2_frauen$session)

# Ende Dataset
### Test people
test_pw2015_t2_ende = pw2015_t2_ende %>%
  filter((session %contains% "XXX") &
         session != "aldt4TnU6Yj8G_EP8PuXXXkLfBmn-03L_NRLpKrR0IRNqKlxIWGQA7ED3yvWF-rZ")
test_pw2015_t2_ende_vector = as.character(test_pw2015_t2_ende$session)

### People based on comments_to_us
comments_pw2015_t2_ende = pw2015_t2_ende %>%
  filter(!is.na(notes_to_us_t2)) %>%
  select(session, notes_to_us_t2)

comments_pw2015_t2_ende = comments_pw2015_t2_ende %>%
  filter(notes_to_us_t2 %contains% "Tanja" | notes_to_us_t2 %contains% "Keine Daten, Testdurchgang!!!"|
         notes_to_us_t2 %contains% "thank the formr monkey\nmany times"|
         notes_to_us_t2 %contains% "Test Silvia"| notes_to_us_t2 %contains% "Test Silvi")
comments_pw2015_t2_ende_vector = as.character(comments_pw2015_t2_ende$session)
```

#### t3
``` {r}
test_pw2015_t3_singles = pw2015_t3_singles %>%
  filter((session %contains% "XXX") &
         session != "aldt4TnU6Yj8G_EP8PuXXXkLfBmn-03L_NRLpKrR0IRNqKlxIWGQA7ED3yvWF-rZ")
test_pw2015_t3_singles_vector = as.character(test_pw2015_t3_singles$session)

test_pw2015_t3_partners = pw2015_t3_partners %>%
  filter((session %contains% "XXX") &
         session != "aldt4TnU6Yj8G_EP8PuXXXkLfBmn-03L_NRLpKrR0IRNqKlxIWGQA7ED3yvWF-rZ")
test_pw2015_t3_partners_vector = as.character(test_pw2015_t3_partners$session)

# Ende
## Test People
test_pw2015_t3_ende = pw2015_t3_ende %>%
  filter((session %contains% "XXX") &
         session != "aldt4TnU6Yj8G_EP8PuXXXkLfBmn-03L_NRLpKrR0IRNqKlxIWGQA7ED3yvWF-rZ")
test_pw2015_t3_ende_vector = as.character(test_pw2015_t3_ende$session)

## Dishonest answers
dishonest_pw2015_t3_ende = pw2015_t3_ende %>%
  filter(F_t3_honesty_check == 2)
dishonest_pw2015_t3_ende_vector = as.character(dishonest_pw2015_t3_ende$session)

## notes to us
comments_pw2015_t3_ende = pw2015_t3_ende %>%
  filter(F_t3_notes_to_us %contains% "thank the formr monkey\nmany times" |
           F_t3_notes_to_us %contains% "Test Silvia" |
           F_t3_notes_to_us == "test" |
           F_t3_notes_to_us %contains% "Test Silvi")
comments_pw2015_t3_ende_vector = as.character(comments_pw2015_t3_ende$session)

```

#### Combine vectors and clean all data sets
```{r}
# Combine all vectors
cleaning_vector = c(test_demographie_vector, test_items_vector, comments_pw2015_items_vector, test_pw2015_t2_vector, test_pw2015_t2_frauen_vector, test_pw2015_t2_ende_vector, comments_pw2015_t2_ende_vector, test_pw2015_t3_singles_vector, test_pw2015_t3_partners_vector,  test_pw2015_t3_ende_vector, dishonest_pw2015_t3_ende_vector, comments_pw2015_t3_ende_vector, NA)

# Remove these people in all datasets and remove missing sessions (NA)
pw2015_demographie =  pw2015_demographie %>%
  filter(!(session %in% cleaning_vector))

pw2015_items =  pw2015_items %>%
  filter(!(session %in% cleaning_vector))  

pw2015_t2 =  pw2015_t2 %>%
  filter(!(session %in% cleaning_vector))
  
pw2015_t2_frauen =  pw2015_t2_frauen %>%
  filter(!(session %in% cleaning_vector))

pw2015_t2_ende =  pw2015_t2_ende %>%
  filter(!(session %in% cleaning_vector))

pw2015_t3_singles =  pw2015_t3_singles %>%
  filter(!(session %in% cleaning_vector))

pw2015_t3_partners =  pw2015_t3_partners %>%
  filter(!(session %in% cleaning_vector))

pw2015_t3_ende =  pw2015_t3_ende %>%
  filter(!(session %in% cleaning_vector))
```


### Check Duplicates
```{r}
dupes_demographie = pw2015_demographie %>% filter(duplicated(session)) #no dupes
dupes_items = pw2015_items %>% filter(duplicated(session)) #no dupes
dupes_t2 = pw2015_t2 %>% filter(duplicated(session))
duplicats = as.character(dupes_t2$session) # remove 3iL0HgCplo from all datasets because of multiple reports of t2 relationships

pw2015_demographie =  pw2015_demographie %>%
  filter(!(session == "3iL0HgCploxC8r7zZxden74E1Z9QmMwkbUw--OsjMp40jStipVrI8aO4Dgc3nngf"))

pw2015_items =  pw2015_items %>%
  filter(!(session == "3iL0HgCploxC8r7zZxden74E1Z9QmMwkbUw--OsjMp40jStipVrI8aO4Dgc3nngf"))  

pw2015_t2 =  pw2015_t2 %>%
  filter(!(session == "3iL0HgCploxC8r7zZxden74E1Z9QmMwkbUw--OsjMp40jStipVrI8aO4Dgc3nngf"))

pw2015_t2_frauen =  pw2015_t2_frauen %>%
  filter(!(session == "3iL0HgCploxC8r7zZxden74E1Z9QmMwkbUw--OsjMp40jStipVrI8aO4Dgc3nngf"))

pw2015_t2_ende =  pw2015_t2_ende %>%
  filter(!(session == "3iL0HgCploxC8r7zZxden74E1Z9QmMwkbUw--OsjMp40jStipVrI8aO4Dgc3nngf"))

pw2015_t3_singles =  pw2015_t3_singles %>%
  filter(!(session == "3iL0HgCploxC8r7zZxden74E1Z9QmMwkbUw--OsjMp40jStipVrI8aO4Dgc3nngf"))

pw2015_t3_partners =  pw2015_t3_partners %>%
  filter(!(session == "3iL0HgCploxC8r7zZxden74E1Z9QmMwkbUw--OsjMp40jStipVrI8aO4Dgc3nngf"))

pw2015_t3_ende =  pw2015_t3_ende %>%
  filter(!(session == "3iL0HgCploxC8r7zZxden74E1Z9QmMwkbUw--OsjMp40jStipVrI8aO4Dgc3nngf"))


dupes_t2_frauen = pw2015_t2_frauen %>% filter(duplicated(session)) # no dupes
dupes_t2_ende = pw2015_t2_ende %>% filter(duplicated(session)) # remove 3iL0HgCplo from all datasets because of multiple reports of t2 relationships



dupes_t3_ende = pw2015_t3_ende %>% filter(duplicated(session))
dupes_t3_ende = pw2015_t3_ende %>% filter(short_session == "ZV_cjcx1fj" | short_session == "BDu_KSxo6P")
dupes_t3_partners = pw2015_t3_partners %>% filter(short_session == "ZV_cjcx1fj" | short_session == "BDu_KSxo6P") # both reported partners for only one date, remove double information from ended dataset, but not from every dataset
pw2015_t3_ende = pw2015_t3_ende %>% filter(!(short_session == "ZV_cjcx1fj" & is.na(ended)) &
                                             !(short_session == "BDu_KSxo6P" & is.na(ended)))

dupes_t3_singles = pw2015_t3_singles %>% filter(duplicated(session)) # no dupes
```

## Data t1: 
### Scale Calculation: 
#### NARQ, NARQ_Adm, NARQ_Riv
```{r}
pw2015_demographie[11:28] <-lapply(pw2015_demographie[11:28], as.numeric)
pw2015_demographie <- pw2015_demographie %>% rowwise () %>%  
  mutate("NARQ_ADM" = mean(c(NARQ_7, NARQ_8, NARQ_1, NARQ_18, NARQ_15, NARQ_16, NARQ_2, NARQ_5, NARQ_3),
                           na.rm = TRUE)) %>% 
  mutate("NARQ_RIV" = mean(c(NARQ_17, NARQ_11, NARQ_13, NARQ_6, NARQ_14, NARQ_9, NARQ_10, NARQ_12, NARQ_4),
                           na.rm = TRUE)) %>%
  mutate("NARQ_scale" = mean(c(NARQ_7, NARQ_8, NARQ_1, NARQ_18, NARQ_15, NARQ_16, NARQ_2, NARQ_5, NARQ_3,
                                NARQ_17, NARQ_11, NARQ_13,  NARQ_6, NARQ_14, NARQ_9, NARQ_10, NARQ_12,
                                NARQ_4 ), na.rm = TRUE)) 
```


#### SOI_R
```{r }
#Item SOI_R_6R umpolen
pw2015_items$SOI_R_6 <- as.integer(recode(pw2015_items$SOI_R_6R, "1" = "7", "2" = "6", "3" = "5", "4" = "4", "5" = "3", "6" = "2", "7" = "1")) 

# scale mean of soi_r
pw2015_items <- pw2015_items %>%  #z-standardisieren, da untersch. Likert-Skalen
  mutate(z_SOI_R_5 = scale(SOI_R_5, center = TRUE, scale = TRUE),
         z_SOI_R_6 = scale(SOI_R_6, center = TRUE, scale = TRUE),
         z_SOI_R_4 = scale(SOI_R_4, center = TRUE, scale = TRUE),
         z_SOI_R_1 = scale(SOI_R_1, center = TRUE, scale = TRUE),
         z_SOI_R_2 = scale(SOI_R_2, center = TRUE, scale = TRUE),
         z_SOI_R_3 = scale(SOI_R_3, center = TRUE, scale = TRUE),
         z_SOI_R_7 = scale(SOI_R_7, center = TRUE, scale = TRUE),
         z_SOI_R_8 = scale(SOI_R_8, center = TRUE, scale = TRUE),
         z_SOI_R_8 = scale(SOI_R_8, center = TRUE, scale = TRUE),
         z_SOI_R_9 = scale(SOI_R_9, center = TRUE, scale = TRUE))

# scale mean of soi_r
pw2015_items <- pw2015_items %>% 
  rowwise() %>%
  mutate(SOI_R_scale = mean(c(z_SOI_R_5, z_SOI_R_6, z_SOI_R_4, z_SOI_R_1, z_SOI_R_2,
                                              z_SOI_R_3, z_SOI_R_7, z_SOI_R_8, z_SOI_R_9), na.rm = TRUE)) %>%
  ungroup()
```

#### Long Term Orientation
``` {r}
pw2015_items$longterm_2 <- as.integer(recode(pw2015_items$longterm_2_R, "1" = "7", "2" = "6", "3" = "5", "4" = "4", "5" = "3", "6" = "2", "7" = "1"))
pw2015_items$longterm_4_M <- as.integer(recode(pw2015_items$longterm_4_R_M, "1" = "7", "2" = "6", "3" = "5", "4" = "4", "5" = "3", "6" = "2", "7" = "1"))
pw2015_items$longterm_4_F <- as.integer(recode(pw2015_items$longterm_4_R_F, "1" = "7", "2" = "6", "3" = "5", "4" = "4", "5" = "3", "6" = "2", "7" = "1"))

pw2015_items <- pw2015_items %>% #z-standardisieren, da untersch. Likert-Skalen
   mutate(z_longterm_1 = scale(longterm_1, center = TRUE, scale = TRUE), z_longterm_2 = scale(longterm_2, center = TRUE, scale = TRUE),
          z_longterm_3_M = scale(longterm_3_M, center = TRUE, scale = TRUE), z_longterm_3_F = scale(longterm_3_F, center = TRUE, scale = TRUE),
          z_longterm_4_M = scale(longterm_4_M, center = TRUE, scale = TRUE), z_longterm_4_F = scale(longterm_4_F, center = TRUE, scale = TRUE),
          z_longterm_5_M = scale(longterm_5_M, center = TRUE, scale = TRUE), z_longterm_5_F = scale(longterm_5_F, center = TRUE, scale = TRUE),
          z_longterm_6_M = scale(longterm_6_M, center = TRUE, scale = TRUE), z_longterm_7_F = scale(longterm_7_F, center = TRUE, scale = TRUE))
 
#Skalenwert Long-Term Orientation
#define items
longterm_items_f <- c("z_longterm_3_F", "z_longterm_4_F", "z_longterm_5_F", "z_longterm_7_F", "z_longterm_1", "z_longterm_2")
longterm_items_m <- c("z_longterm_3_M", "z_longterm_4_M", "z_longterm_5_M", "z_longterm_6_M", "z_longterm_1", "z_longterm_2")

#aggregate to scale
pw2015_items = pw2015_items %>%
  mutate(longterm_f =  rowMeans(pw2015_items[,longterm_items_f], na.rm = T),
         longterm_f = ifelse(is.na(z_longterm_3_F), NA, longterm_f),
         longterm_m =  rowMeans(pw2015_items[,longterm_items_m], na.rm = T),
         longterm_m = ifelse(is.na(z_longterm_3_M), NA, longterm_m))
         
#crosstabs(~ is.na(longterm_f) + is.na(longterm_m), data = pw2015_items)
```

###BFI items: scale
```{r}
##recode variable ending with R
pw2015_demographie = pw2015_demographie %>%
  mutate(BFIK_open_5 = as.numeric(recode(BFIK_open_5R,
                                         "0" = "5", "1" = "4", "2" = "3", "3" = "2", "4" = "1", "5" = "0")),
         BFIK_agree_1 = as.numeric(recode(BFIK_agree_1R,
                                          "0" = "5", "1" = "4", "2" = "3", "3" = "2", "4" = "1", "5" = "0")),
         BFIK_neuro_2 = as.numeric(recode(BFIK_neuro_2R,
                                          "0" = "5", "1" = "4", "2" = "3", "3" = "2", "4" = "1", "5" = "0")),
         BFIK_consc_2 = as.numeric(recode(BFIK_consc_2R,
                                          "0" = "5", "1" = "4", "2" = "3", "3" = "2", "4" = "1", "5" = "0")),
         BFIK_agree_3 = as.numeric(recode(BFIK_agree_3R,
                                          "0" = "5", "1" = "4", "2" = "3", "3" = "2", "4" = "1", "5" = "0")),
         BFIK_extra_3 = as.numeric(recode(BFIK_extra_3R,
                                          "0" = "5", "1" = "4", "2" = "3", "3" = "2", "4" = "1", "5" = "0")),
         BFIK_extra_1 = as.numeric(recode(BFIK_extra_1R,
                                          "0" = "5", "1" = "4", "2" = "3", "3" = "2", "4" = "1", "5" = "0")),
         )

## form scale
pw2015_demographie = pw2015_demographie %>%
  mutate(BFIK_open = mean(c(BFIK_open_1, BFIK_open_2, BFIK_open_3, BFIK_open_4, BFIK_open_5), na.rm = T),
         BFIK_consc = mean(c(BFIK_consc_1, BFIK_consc_2, BFIK_consc_3, BFIK_consc_4), na.rm = T),
         BFIK_extra = mean(c(BFIK_extra_1, BFIK_extra_2, BFIK_extra_3, BFIK_extra_4), na.rm = T),
         BFIK_agree = mean(c(BFIK_agree_1, BFIK_agree_2, BFIK_agree_3, BFIK_agree_4), na.rm = T),
         BFIK_neuro = mean(c(BFIK_neuro_1, BFIK_neuro_2, BFIK_neuro_3, BFIK_neuro_4)), na.rm = T)

## check scales
### overall
pw2015_demographie %>% select(BFIK_open, BFIK_consc, BFIK_extra, BFIK_agree, BFIK_neuro) %>% apa.cor.table()

### open
pw2015_demographie %>% select(BFIK_open, BFIK_open_1, BFIK_open_2, BFIK_open_3, BFIK_open_4, BFIK_open_5) %>%
  apa.cor.table()

### consc
pw2015_demographie %>% select(BFIK_consc, BFIK_consc_1, BFIK_consc_2, BFIK_consc_3, BFIK_consc_4) %>%
  apa.cor.table()

### extra
pw2015_demographie %>% select(BFIK_extra, BFIK_extra_1, BFIK_extra_2, BFIK_extra_3, BFIK_extra_4) %>%
  apa.cor.table()

### agree
pw2015_demographie %>% select(BFIK_agree, BFIK_agree_1, BFIK_agree_2, BFIK_agree_3, BFIK_agree_4) %>%
  apa.cor.table()

### neuro
pw2015_demographie %>% select(BFIK_neuro, BFIK_neuro_1, BFIK_neuro_2, BFIK_neuro_3, BFIK_neuro_4) %>%
  apa.cor.table()
```

###Time since last relationship as date Variable --> create Variable "previous_relationship_ended"
```{r}
pw2015_items <-  pw2015_items %>%
  mutate(created = as.Date(created))

# For people who never had a relationship we need their age to calculate backwards how long they have been single
pw2015_items = left_join(pw2015_items, pw2015_demographie %>% select(session, age), by = "session")

#################################################################### 
##########Is it accurate to calculate with age, if always single?
pw2015_items = pw2015_items %>%
  mutate(end_relationship_years = ifelse(number_relationships == 0, age, end_relationship_years))
####################################################################



pw2015_items <- pw2015_items %>%
  mutate(previous_relationship_ended_days = end_relationship_years * 365 + 
                                            end_relationship_months * 30 + 
                                            end_relationship_weeks * 7,
         previous_relationship_ended_days = ifelse(is.na(previous_relationship_ended_days),
                                                   end_relationship_years * 365 + 
                                                     end_relationship_months * 30,
                                                   previous_relationship_ended_days),
         previous_relationship_ended_days = ifelse(is.na(previous_relationship_ended_days),
                                                   end_relationship_years * 365,
                                                   previous_relationship_ended_days),
         previous_relationship_ended = created - previous_relationship_ended_days)


#################### Why still so many participants with without "previous_relationship_ended"?? 
crosstabs(~ number_relationships + is.na(previous_relationship_ended), data = pw2015_items)

pw2015_items <- pw2015_items %>% select(-previous_relationship_ended_days,  -end_relationship_years,
                                        -end_relationship_months, -end_relationship_weeks)
```



## Data t2
### Join male-interested (pw2015_t2) and female-interested (pw2015_t2_frauen) dataset
``` {r }
pw2015_t2$session = as.character(pw2015_t2$session)
pw2015_t2_frauen$session = as.character(pw2015_t2_frauen$session)
pw2015_t2$short_session = as.character(pw2015_t2$short_session)
pw2015_t2_frauen$short_session = as.character(pw2015_t2_frauen$short_session)

# pw2015_t2_relationship <- bind_rows(pw2015_t2, pw2015_t2_frauen) @Sarah: check if we can change to this version
pw2015_t2_relationship <- left_join(pw2015_t2, pw2015_t2_frauen, by=c("session", "short_session"), suffix = c(".m", ".f"))
```

### Relationship information t2

```{r}
pw2015_t2_relationship <- pw2015_t2_relationship %>%
  mutate(
  "t2_relationship" = derivedVariable(
            "0" = (relationship_f == 1 & relationship_m == 1) | (relationship_f == 1 & is.na(relationship_m) |  
                          (is.na(relationship_f) & relationship_m == 1)),
            "1" = (relationship_f == 2 | relationship_m == 2),
            "2" = (relationship_f == 3 | relationship_m == 3), .method = "first", .default = 0))

```


### T2: Create start (date) variable of current relationships 
```{r}
pw2015_t2_relationship = pw2015_t2_relationship %>%
  mutate(created.m = as.Date(created.m),
         created.f = as.Date(created.f))


pw2015_t2_relationship <- pw2015_t2_relationship %>%
  mutate(relationship1_start_days_m = relationship1_start_months_m * 30 + relationship1_start_weeks_m * 7,
         relationship1_start_m = created.m - relationship1_start_days_m, 
         relationship1_start_days_f = relationship1_start_months_f * 30 + relationship1_start_weeks_f * 7,
         relationship1_start_f = created.f - relationship1_start_days_f)

pw2015_t2_relationship <- pw2015_t2_relationship %>%
  mutate(relationship2_start_days_m = relationship2_start_months_m * 30 + relationship2_start_weeks_m * 7,
         relationship2_start_m = created.m - relationship2_start_days_m, 
         relationship2_start_days_f = relationship2_start_months_f * 30 + relationship2_start_weeks_f * 7,
         relationship2_start_f = created.f - relationship2_start_days_f)

pw2015_t2_relationship <- pw2015_t2_relationship %>%
  mutate(relationship3_start_days_m = relationship3_start_months_m * 30 + relationship3_start_weeks_m * 7,
         relationship3_start_m = created.m - relationship3_start_days_m, 
         relationship3_start_days_f = relationship3_start_months_f * 30 + relationship3_start_weeks_f * 7,
         relationship3_start_f = created.f - relationship3_start_days_f)

#filter unnecessary start variables
pw2015_t2_relationship <- pw2015_t2_relationship %>% select(-relationship1_start_months_m, -relationship2_start_months_m,
                       -relationship3_start_months_m, -relationship1_start_weeks_m, -relationship2_start_weeks_m,    
                       -relationship3_start_weeks_m, -relationship1_start_days_m, -relationship2_start_days_m,
                       -relationship3_start_days_m, -relationship1_start_months_f, -relationship2_start_months_f,
                       -relationship3_start_months_f, -relationship1_start_weeks_f, -relationship2_start_weeks_f,    
                       -relationship3_start_weeks_f, -relationship1_start_days_f, -relationship2_start_days_f,
                       -relationship3_start_days_f)
```

### Divide current relationships depending on sex 

#### Relationship 1
```{r}
pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "relationship4_kind" = derivedVariable(
     "1" = (!is.na(relationship1_kind_m) & !is.na(relationship1_kind_f)),.method = "first", .default = 0))
pw2015_t2_relationship$relationship4_kind[pw2015_t2_relationship$relationship4_kind == 1] <- pw2015_t2_relationship$relationship1_kind_f[pw2015_t2_relationship$relationship4_kind == 1]
pw2015_t2_relationship$relationship4_kind[pw2015_t2_relationship$relationship4_kind == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "relationship4_start" = derivedVariable(
     "1" = (!is.na(relationship1_start_m) & !is.na(relationship1_start_f)),.method = "first", .default = 0))
pw2015_t2_relationship$relationship4_start[pw2015_t2_relationship$relationship4_start == 1] <- pw2015_t2_relationship$relationship1_start_f[pw2015_t2_relationship$relationship4_start == 1]
pw2015_t2_relationship$relationship4_start[pw2015_t2_relationship$relationship4_start == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "relationship4_name" = derivedVariable(
     "1" = (!is.na(relationship1_m_name) & !is.na(relationship1_f_name)),.method = "first", .default = 0))
pw2015_t2_relationship$relationship4_name[pw2015_t2_relationship$relationship4_name == 1] <- pw2015_t2_relationship$relationship1_f_name[pw2015_t2_relationship$relationship4_name == 1]
pw2015_t2_relationship$relationship4_name[pw2015_t2_relationship$relationship4_name == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "relationship4_name_new" = derivedVariable(
     "1" = (!is.na(relationship1_name_m_new) & !is.na(relationship1_name_f_new)),.method = "first", .default = 0))
pw2015_t2_relationship$relationship4_name_new[pw2015_t2_relationship$relationship4_name_new == 1] <- pw2015_t2_relationship$relationship1_name_f_new[pw2015_t2_relationship$relationship4_name_new == 1]
pw2015_t2_relationship$relationship4_name_new[pw2015_t2_relationship$relationship4_name_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "relationship4_age_new" = derivedVariable(
     "1" = (!is.na(relationship1_age_m_new) & !is.na(relationship1_age_f_new)),.method = "first", .default = 0))
pw2015_t2_relationship$relationship4_age_new[pw2015_t2_relationship$relationship4_age_new == 1] <- pw2015_t2_relationship$relationship1_age_f_new[pw2015_t2_relationship$relationship4_age_new == 1]
pw2015_t2_relationship$relationship4_age_new[pw2015_t2_relationship$relationship4_age_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "relationship4_height_new" = derivedVariable(
     "1" = (!is.na(relationship1_height_m_new) & !is.na(relationship1_height_f_new)),.method = "first", .default = 0))
pw2015_t2_relationship$relationship4_height_new[pw2015_t2_relationship$relationship4_height_new == 1] <- pw2015_t2_relationship$relationship1_height_f_new[pw2015_t2_relationship$relationship4_height_new == 1]
pw2015_t2_relationship$relationship4_height_new[pw2015_t2_relationship$relationship4_height_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "relationship4_relationship_new" = derivedVariable(
     "1" = (!is.na(relationship1_relationship_m_new) & !is.na(relationship1_relationship_f_new)),.method = "first", .default = 0))
pw2015_t2_relationship$relationship4_relationship_new[pw2015_t2_relationship$relationship4_relationship_new == 1] <- pw2015_t2_relationship$relationship1_relationship_f_new[pw2015_t2_relationship$relationship4_relationship_new == 1]
pw2015_t2_relationship$relationship4_relationship_new[pw2015_t2_relationship$relationship4_relationship_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "relationship4_duration_new" = derivedVariable(
     "1" = (!is.na(relationship1_duration_m_new) & !is.na(relationship1_duration_f_new)),.method = "first", .default = 0))
pw2015_t2_relationship$relationship4_duration_new[pw2015_t2_relationship$relationship4_duration_new == 1] <- pw2015_t2_relationship$relationship1_duration_f_new[pw2015_t2_relationship$relationship4_duration_new == 1]
pw2015_t2_relationship$relationship4_duration_new[pw2015_t2_relationship$relationship4_duration_new == 0] <- NA

##not necessary? -> created additional start variable
#pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
 # "relationship4_start_months" = derivedVariable(
  #   "1" = (!is.na(relationship1_start_months_m) & !is.na(relationship1_start_months_f)) ,.method = "first", .default = 0))
#pw2015_t2_relationship$relationship4_start_months[pw2015_t2_relationship$relationship4_start_months == 1] <- #pw2015_t2_relationship$relationship1_start_months_f[pw2015_t2_relationship$relationship4_start_months == 1]
#pw2015_t2_relationship$relationship4_start_months[pw2015_t2_relationship$relationship4_start_months == 0] <- NA

##not necessary? -> created additional start variable
#pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
 # "relationship4_start_weeks" = derivedVariable(
  #   "1" = (!is.na(relationship1_start_weeks_m) & !is.na(relationship1_start_weeks_f)),.method = "first", .default = 0))
#pw2015_t2_relationship$relationship4_start_weeks[pw2015_t2_relationship$relationship4_start_weeks == 1] <- #pw2015_t2_relationship$relationship1_start_weeks_f[pw2015_t2_relationship$relationship4_start_weeks == 1]
#pw2015_t2_relationship$relationship4_start_weeks[pw2015_t2_relationship$relationship4_start_weeks == 0] <- NA
```

#### Relationship 2
```{r}
pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "relationship5_kind" = derivedVariable(
     "1" = (!is.na(relationship2_kind_m) & !is.na(relationship2_kind_f)),.method = "first", .default = 0))
pw2015_t2_relationship$relationship5_kind[pw2015_t2_relationship$relationship5_kind == 1] <- pw2015_t2_relationship$relationship2_kind_f[pw2015_t2_relationship$relationship5_kind == 1]
pw2015_t2_relationship$relationship5_kind[pw2015_t2_relationship$relationship5_kind == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "relationship5_start" = derivedVariable(
     "1" = (!is.na(relationship2_start_m) & !is.na(relationship2_start_f)),.method = "first", .default = 0))
pw2015_t2_relationship$relationship5_start[pw2015_t2_relationship$relationship5_start == 1] <- pw2015_t2_relationship$relationship2_start_f[pw2015_t2_relationship$relationship5_start == 1]
pw2015_t2_relationship$relationship5_start[pw2015_t2_relationship$relationship5_start == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "relationship5_name" = derivedVariable(
     "1" = (!is.na(relationship2_m_name) & !is.na(relationship2_f_name)),.method = "first", .default = 0))
pw2015_t2_relationship$relationship5_name[pw2015_t2_relationship$relationship5_name == 1] <- pw2015_t2_relationship$relationship2_f_name[pw2015_t2_relationship$relationship5_name == 1]
pw2015_t2_relationship$relationship5_name[pw2015_t2_relationship$relationship2_name == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "relationship5_name_new" = derivedVariable(
     "1" = (!is.na(relationship2_name_m_new) & !is.na(relationship2_name_f_new)),.method = "first", .default = 0))
pw2015_t2_relationship$relationship5_name_new[pw2015_t2_relationship$relationship5_name_new == 1] <- pw2015_t2_relationship$relationship2_name_f_new[pw2015_t2_relationship$relationship5_name_new == 1]
pw2015_t2_relationship$relationship5_name_new[pw2015_t2_relationship$relationship5_name_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "relationship5_age_new" = derivedVariable(
     "1" = (!is.na(relationship2_age_m_new) & !is.na(relationship2_age_f_new)),.method = "first", .default = 0))
pw2015_t2_relationship$relationship5_age_new[pw2015_t2_relationship$relationship5_age_new == 1] <- pw2015_t2_relationship$relationship2_age_f_new[pw2015_t2_relationship$relationship5_age_new == 1]
pw2015_t2_relationship$relationship5_age_new[pw2015_t2_relationship$relationship5_age_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "relationship5_height_new" = derivedVariable(
     "1" = (!is.na(relationship2_height_m_new) & !is.na(relationship2_height_f_new)),.method = "first", .default = 0))
pw2015_t2_relationship$relationship4_height_new[pw2015_t2_relationship$relationship5_height_new == 1] <- pw2015_t2_relationship$relationship2_height_f_new[pw2015_t2_relationship$relationship5_height_new == 1]
pw2015_t2_relationship$relationship5_height_new[pw2015_t2_relationship$relationship5_height_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "relationship5_relationship_new" = derivedVariable(
     "1" = (!is.na(relationship2_relationship_m_new) & !is.na(relationship2_relationship_f_new)),.method = "first", .default = 0))
pw2015_t2_relationship$relationship5_relationship_new[pw2015_t2_relationship$relationship5_relationship_new == 1] <- pw2015_t2_relationship$relationship2_relationship_f_new[pw2015_t2_relationship$relationship5_relationship_new == 1]
pw2015_t2_relationship$relationship5_relationship_new[pw2015_t2_relationship$relationship5_relationship_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "relationship5_duration_new" = derivedVariable(
     "1" = (!is.na(relationship2_duration_m_new) & !is.na(relationship2_duration_f_new)),.method = "first", .default = 0))
pw2015_t2_relationship$relationship5_duration_new[pw2015_t2_relationship$relationship5_duration_new == 1] <- pw2015_t2_relationship$relationship2_duration_f_new[pw2015_t2_relationship$relationship5_duration_new == 1]
pw2015_t2_relationship$relationship5_duration_new[pw2015_t2_relationship$relationship5_duration_new == 0] <- NA

##not necessary? -> created additional start variable
#pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
 # "relationship5_start_months" = derivedVariable(
  #   "1" = (!is.na(relationship2_start_months_m) & !is.na(relationship2_start_months_f)) ,.method = "first", .default = 0))
#pw2015_t2_relationship$relationship5_start_months[pw2015_t2_relationship$relationship5_start_months == 1] <- #pw2015_t2_relationship$relationship2_start_months_f[pw2015_t2_relationship$relationship5_start_months == 1]
#pw2015_t2_relationship$relationship5_start_months[pw2015_t2_relationship$relationship5_start_months == 0] <- NA

##not necessary? -> created additional start variable
#pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
 # "relationship5_start_weeks" = derivedVariable(
  #   "1" = (!is.na(relationship2_start_weeks_m) & !is.na(relationship2_start_weeks_f)),.method = "first", .default = 0))
#pw2015_t2_relationship$relationship5_start_weeks[pw2015_t2_relationship$relationship5_start_weeks == 1] <- #pw2015_t2_relationship$relationship2_start_weeks_f[pw2015_t2_relationship$relationship5_start_weeks == 1]
#pw2015_t2_relationship$relationship5_start_weeks[pw2015_t2_relationship$relationship5_start_weeks == 0] <- NA
```

#### Relationship 3
```{r}
pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "relationship6_kind" = derivedVariable(
     "1" = (!is.na(relationship3_kind_m) & !is.na(relationship3_kind_f)),.method = "first", .default = 0))
pw2015_t2_relationship$relationship6_kind[pw2015_t2_relationship$relationship6_kind == 1] <- pw2015_t2_relationship$relationship3_kind_f[pw2015_t2_relationship$relationship6_kind == 1]
pw2015_t2_relationship$relationship6_kind[pw2015_t2_relationship$relationship6_kind == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "relationship6_start" = derivedVariable(
     "1" = (!is.na(relationship3_start_m) & !is.na(relationship3_start_f)),.method = "first", .default = 0))
pw2015_t2_relationship$relationship6_start[pw2015_t2_relationship$relationship6_start == 1] <- pw2015_t2_relationship$relationship3_start_f[pw2015_t2_relationship$relationship6_start == 1]
pw2015_t2_relationship$relationship6_start[pw2015_t2_relationship$relationship6_start == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "relationship6_name" = derivedVariable(
     "1" = (!is.na(relationship3_m_name) & !is.na(relationship3_f_name)),.method = "first", .default = 0))
pw2015_t2_relationship$relationship6_name[pw2015_t2_relationship$relationship6_name == 1] <- pw2015_t2_relationship$relationship3_f_name[pw2015_t2_relationship$relationship6_name == 1]
pw2015_t2_relationship$relationship6_name[pw2015_t2_relationship$relationship3_name == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "relationship6_name_new" = derivedVariable(
     "1" = (!is.na(relationship3_name_m_new) & !is.na(relationship3_name_f_new)),.method = "first", .default = 0))
pw2015_t2_relationship$relationship6_name_new[pw2015_t2_relationship$relationship6_name_new == 1] <- pw2015_t2_relationship$relationship3_name_f_new[pw2015_t2_relationship$relationship6_name_new == 1]
pw2015_t2_relationship$relationship6_name_new[pw2015_t2_relationship$relationship6_name_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "relationship6_age_new" = derivedVariable(
     "1" = (!is.na(relationship3_age_m_new) & !is.na(relationship3_age_f_new)),.method = "first", .default = 0))
pw2015_t2_relationship$relationship6_age_new[pw2015_t2_relationship$relationship6_age_new == 1] <- pw2015_t2_relationship$relationship3_age_f_new[pw2015_t2_relationship$relationship6_age_new == 1]
pw2015_t2_relationship$relationship6_age_new[pw2015_t2_relationship$relationship6_age_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "relationship6_height_new" = derivedVariable(
     "1" = (!is.na(relationship3_height_m_new) & !is.na(relationship3_height_f_new)),.method = "first", .default = 0))
pw2015_t2_relationship$relationship6_height_new[pw2015_t2_relationship$relationship6_height_new == 1] <- pw2015_t2_relationship$relationship3_height_f_new[pw2015_t2_relationship$relationship6_height_new == 1]
pw2015_t2_relationship$relationship6_height_new[pw2015_t2_relationship$relationship6_height_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "relationship6_relationship_new" = derivedVariable(
     "1" = (!is.na(relationship3_relationship_m_new) & !is.na(relationship3_relationship_f_new)),.method = "first", .default = 0))
pw2015_t2_relationship$relationship6_relationship_new[pw2015_t2_relationship$relationship6_relationship_new == 1] <- pw2015_t2_relationship$relationship3_relationship_f_new[pw2015_t2_relationship$relationship6_relationship_new == 1]
pw2015_t2_relationship$relationship6_relationship_new[pw2015_t2_relationship$relationship6_relationship_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "relationship6_duration_new" = derivedVariable(
     "1" = (!is.na(relationship3_duration_m_new) & !is.na(relationship3_duration_f_new)),.method = "first", .default = 0))
pw2015_t2_relationship$relationship6_duration_new[pw2015_t2_relationship$relationship6_duration_new == 1] <- pw2015_t2_relationship$relationship3_duration_f_new[pw2015_t2_relationship$relationship6_duration_new == 1]
pw2015_t2_relationship$relationship6_duration_new[pw2015_t2_relationship$relationship6_duration_new == 0] <- NA

##not necessary? -> created additional start variable
#pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
 # "relationship6_start_months" = derivedVariable(
  #   "1" = (!is.na(relationship3_start_months_m) & !is.na(relationship3_start_months_f)) ,.method = "first", .default = 0))
#pw2015_t2_relationship$relationship6_start_months[pw2015_t2_relationship$relationship6_start_months == 1] <- #pw2015_t2_relationship$relationship3_start_months_f[pw2015_t2_relationship$relationship6_start_months == 1]
#pw2015_t2_relationship$relationship6_start_months[pw2015_t2_relationship$relationship6_start_months == 0] <- NA

##not necessary? -> created additional start variable
#pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
 # "relationship6_start_weeks" = derivedVariable(
  #   "1" = (!is.na(relationship1_start_weeks_m) & !is.na(relationship1_start_weeks_f)),.method = "first", .default = 0))
#pw2015_t2_relationship$relationship6_start_weeks[pw2015_t2_relationship$relationship6_start_weeks == 1] <- #pw2015_t2_relationship$relationship3_start_weeks_f[pw2015_t2_relationship$relationship6_start_weeks == 1]
#pw2015_t2_relationship$relationship6_start_weeks[pw2015_t2_relationship$relationship6_start_weeks == 0] <- NA
```



### Unite Variables
```{r}
pw2015_t2_relationship <- pw2015_t2_relationship %>% unite(t2_relationship_multiple, relationship_m_multiple, relationship_f_multiple)


#Relationship 1
pw2015_t2_relationship <- pw2015_t2_relationship %>% unite(relationship1_kind, relationship1_kind_m, relationship1_kind_f) %>%
      unite(relationship1_start, relationship1_start_m, relationship1_start_f) %>% 
      unite(relationship1_name, relationship1_m_name, relationship1_f_name) %>% 
      unite(relationship1_name_new, relationship1_name_m_new, relationship1_name_f_new) %>% 
      unite(relationship1_age_new, relationship1_age_m_new, relationship1_age_f_new) %>% 
      unite(relationship1_height_new, relationship1_height_m_new, relationship1_height_f_new) %>%
      unite(relationship1_relationship_new, relationship1_relationship_m_new, relationship1_relationship_f_new) %>%
      unite(relationship1_duration_new, relationship1_duration_m_new, relationship1_duration_f_new)

#Relationship 2
pw2015_t2_relationship <- pw2015_t2_relationship %>% unite(relationship2_kind, relationship2_kind_m, relationship2_kind_f) %>%
      unite(relationship2_start, relationship2_start_m, relationship2_start_f) %>% 
      unite(relationship2_name, relationship2_m_name, relationship2_f_name) %>% 
      unite(relationship2_name_new, relationship2_name_m_new, relationship2_name_f_new) %>% 
      unite(relationship2_age_new, relationship2_age_m_new, relationship2_age_f_new) %>% 
      unite(relationship2_height_new, relationship2_height_m_new, relationship2_height_f_new) %>%
      unite(relationship2_relationship_new, relationship2_relationship_m_new, relationship2_relationship_f_new) %>%
      unite(relationship2_duration_new, relationship2_duration_m_new, relationship2_duration_f_new)

#Relationship 3
pw2015_t2_relationship <- pw2015_t2_relationship %>% unite(relationship3_kind, relationship3_kind_m, relationship3_kind_f) %>%
      unite(relationship3_start, relationship3_start_m, relationship3_start_f) %>% 
      unite(relationship3_name, relationship3_m_name, relationship3_f_name) %>% 
      unite(relationship3_name_new, relationship3_name_m_new, relationship3_name_f_new) %>% 
      unite(relationship3_age_new, relationship3_age_m_new, relationship3_age_f_new) %>% 
      unite(relationship3_height_new, relationship3_height_m_new, relationship3_height_f_new) %>%
      unite(relationship3_relationship_new, relationship3_relationship_m_new, relationship3_relationship_f_new) %>%
      unite(relationship3_duration_new, relationship3_duration_m_new, relationship3_duration_f_new) 

```

### Divide variables, necessary when VPs had both male and female partners

#### Relationship 1
```{r}

pw2015_t2_relationship$relationship1_kind <- sub('_NA*','', pw2015_t2_relationship$relationship1_kind)
pw2015_t2_relationship$relationship1_kind <- sub('NA_*','', pw2015_t2_relationship$relationship1_kind)
pw2015_t2_relationship$relationship1_kind[pw2015_t2_relationship$relationship1_kind==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', relationship1_kind)) %>% separate(relationship1_kind, c("relationship1_kind","relationship4_kind"), sep = "_")

pw2015_t2_relationship$relationship1_start <- sub('_NA*','', pw2015_t2_relationship$relationship1_start)
pw2015_t2_relationship$relationship1_start <- sub('NA_*','', pw2015_t2_relationship$relationship1_start)
pw2015_t2_relationship$relationship1_start[pw2015_t2_relationship$relationship1_start==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', relationship1_start)) %>% separate(relationship1_start, c("relationship1_start","relationship4_start"), sep = "_")

pw2015_t2_relationship$relationship1_name <- sub('_NA*','', pw2015_t2_relationship$relationship1_name)
pw2015_t2_relationship$relationship1_name <- sub('NA_*','', pw2015_t2_relationship$relationship1_name)
pw2015_t2_relationship$relationship1_name[pw2015_t2_relationship$relationship1_name==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', relationship1_name)) %>% separate(relationship1_name, c("relationship1_name","relationship4_name"), sep = "_")

pw2015_t2_relationship$relationship1_name_new <- sub('_NA*','', pw2015_t2_relationship$relationship1_name_new)
pw2015_t2_relationship$relationship1_name_new <- sub('NA_*','', pw2015_t2_relationship$relationship1_name_new)
pw2015_t2_relationship$relationship1_name_new[pw2015_t2_relationship$relationship1_name_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', relationship1_name_new)) %>% separate(relationship1_name_new, c("relationship1_name_new","relationship4_name_new"), sep = "_")

pw2015_t2_relationship$relationship1_age_new <- sub('_NA*','', pw2015_t2_relationship$relationship1_age_new)
pw2015_t2_relationship$relationship1_age_new <- sub('NA_*','', pw2015_t2_relationship$relationship1_age_new)
pw2015_t2_relationship$relationship1_age_new[pw2015_t2_relationship$relationship1_age_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', relationship1_age_new)) %>% separate(relationship1_age_new, c("relationship1_age_new","relationship4_age_new"), sep = "_")

pw2015_t2_relationship$relationship1_height_new <- sub('_NA*','', pw2015_t2_relationship$relationship1_height_new)
pw2015_t2_relationship$relationship1_height_new <- sub('NA_*','', pw2015_t2_relationship$relationship1_height_new)
pw2015_t2_relationship$relationship1_height_new[pw2015_t2_relationship$relationship1_height_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', relationship1_height_new)) %>% separate(relationship1_height_new, c("relationship1_height_new","relationship4_height_new"), sep = "_")

pw2015_t2_relationship$relationship1_relationship_new <- sub('_NA*','', pw2015_t2_relationship$relationship1_relationship_new)
pw2015_t2_relationship$relationship1_relationship_new <- sub('NA_*','', pw2015_t2_relationship$relationship1_relationship_new)
pw2015_t2_relationship$relationship1_relationship_new[pw2015_t2_relationship$relationship1_relationship_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', relationship1_relationship_new)) %>% separate(relationship1_relationship_new, c("relationship1_relationship_new","relationship4_relationship_new"), sep = "_")

pw2015_t2_relationship$relationship1_duration_new <- sub('_NA*','', pw2015_t2_relationship$relationship1_duration_new)
pw2015_t2_relationship$relationship1_duration_new <- sub('NA_*','', pw2015_t2_relationship$relationship1_duration_new)
pw2015_t2_relationship$relationship1_duration_new[pw2015_t2_relationship$relationship1_duration_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', relationship1_duration_new)) %>% separate(relationship1_duration_new, c("relationship1_duration_new","relationship4_duration_new"), sep = "_")
```

Relationship 2
```{r}
pw2015_t2_relationship$relationship2_kind <- sub('_NA*','', pw2015_t2_relationship$relationship2_kind)
pw2015_t2_relationship$relationship2_kind <- sub('NA_*','', pw2015_t2_relationship$relationship2_kind)
pw2015_t2_relationship$relationship2_kind[pw2015_t2_relationship$relationship2_kind==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', relationship2_kind)) %>% separate(relationship2_kind, c("relationship2_kind","relationship5_kind"), sep = "_")

pw2015_t2_relationship$relationship2_start <- sub('_NA*','', pw2015_t2_relationship$relationship2_start)
pw2015_t2_relationship$relationship2_start <- sub('NA_*','', pw2015_t2_relationship$relationship2_start)
pw2015_t2_relationship$relationship2_start[pw2015_t2_relationship$relationship2_start==""]  <- NA 
z <- pw2015_t2_relationship %>% group_by(grepl('_', relationship2_start)) %>% separate(relationship2_start, c("relationship2_start","relationship5_start"), sep = "_")

pw2015_t2_relationship$relationship2_name <- sub('_NA*','', pw2015_t2_relationship$relationship2_name)
pw2015_t2_relationship$relationship2_name <- sub('NA_*','', pw2015_t2_relationship$relationship2_name)
pw2015_t2_relationship$relationship2_name[pw2015_t2_relationship$relationship2_name==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', relationship2_name)) %>% separate(relationship2_name, c("relationship2_name","relationship5_name"), sep = "_")

pw2015_t2_relationship$relationship2_name_new <- sub('_NA*','', pw2015_t2_relationship$relationship2_name_new)
pw2015_t2_relationship$relationship2_name_new <- sub('NA_*','', pw2015_t2_relationship$relationship2_name_new)
pw2015_t2_relationship$relationship2_name_new[pw2015_t2_relationship$relationship2_name_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', relationship2_name_new)) %>% separate(relationship2_name_new, c("relationship2_name_new","relationship5_name_new"), sep = "_")

pw2015_t2_relationship$relationship2_age_new <- sub('_NA*','', pw2015_t2_relationship$relationship2_age_new)
pw2015_t2_relationship$relationship2_age_new <- sub('NA_*','', pw2015_t2_relationship$relationship2_age_new)
pw2015_t2_relationship$relationship2_age_new[pw2015_t2_relationship$relationship2_age_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', relationship2_age_new)) %>% separate(relationship2_age_new, c("relationship2_age_new","relationship5_age_new"), sep = "_")

pw2015_t2_relationship$relationship2_height_new <- sub('_NA*','', pw2015_t2_relationship$relationship2_height_new)
pw2015_t2_relationship$relationship2_height_new <- sub('NA_*','', pw2015_t2_relationship$relationship2_height_new)
pw2015_t2_relationship$relationship2_height_new[pw2015_t2_relationship$relationship2_height_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', relationship2_height_new)) %>% separate(relationship2_height_new, c("relationship2_height_new","relationship5_height_new"), sep = "_")

pw2015_t2_relationship$relationship2_relationship_new <- sub('_NA*','', pw2015_t2_relationship$relationship2_relationship_new)
pw2015_t2_relationship$relationship2_relationship_new <- sub('NA_*','', pw2015_t2_relationship$relationship2_relationship_new)
pw2015_t2_relationship$relationship2_relationship_new[pw2015_t2_relationship$relationship2_relationship_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', relationship2_relationship_new)) %>% separate(relationship2_relationship_new, c("relationship2_relationship_new","relationship5_relationship_new"), sep = "_")

pw2015_t2_relationship$relationship2_duration_new <- sub('_NA*','', pw2015_t2_relationship$relationship2_duration_new)
pw2015_t2_relationship$relationship2_duration_new <- sub('NA_*','', pw2015_t2_relationship$relationship2_duration_new)
pw2015_t2_relationship$relationship2_duration_new[pw2015_t2_relationship$relationship2_duration_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', relationship2_duration_new)) %>% separate(relationship2_duration_new, c("relationship2_duration_new","relationship5_duration_new"), sep = "_")
```

Relationship 3
```{r}
pw2015_t2_relationship$relationship3_kind <- sub('_NA*','', pw2015_t2_relationship$relationship3_kind)
pw2015_t2_relationship$relationship3_kind <- sub('NA_*','', pw2015_t2_relationship$relationship3_kind)
pw2015_t2_relationship$relationship3_kind[pw2015_t2_relationship$relationship3_kind==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', relationship3_kind)) %>% separate(relationship3_kind, c("relationship3_kind","relationship6_kind"), sep = "_")

pw2015_t2_relationship$relationship3_start <- sub('_NA*','', pw2015_t2_relationship$relationship3_start)
pw2015_t2_relationship$relationship3_start <- sub('NA_*','', pw2015_t2_relationship$relationship3_start)
pw2015_t2_relationship$relationship3_start[pw2015_t2_relationship$relationship3_start==""]  <- NA 
z <- pw2015_t2_relationship %>% group_by(grepl('_', relationship3_start)) %>% separate(relationship3_start, c("relationship3_start","relationship6_start"), sep = "_")

pw2015_t2_relationship$relationship3_name <- sub('_NA*','', pw2015_t2_relationship$relationship3_name)
pw2015_t2_relationship$relationship3_name <- sub('NA_*','', pw2015_t2_relationship$relationship3_name)
pw2015_t2_relationship$relationship3_name[pw2015_t2_relationship$relationship3_name==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', relationship3_name)) %>% separate(relationship3_name, c("relationship3_name","relationship6_name"), sep = "_")

pw2015_t2_relationship$relationship3_name_new <- sub('_NA*','', pw2015_t2_relationship$relationship3_name_new)
pw2015_t2_relationship$relationship3_name_new <- sub('NA_*','', pw2015_t2_relationship$relationship3_name_new)
pw2015_t2_relationship$relationship3_name_new[pw2015_t2_relationship$relationship3_name_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', relationship3_name_new)) %>% separate(relationship3_name_new, c("relationship3_name_new","relationship6_name_new"), sep = "_")

pw2015_t2_relationship$relationship3_age_new <- sub('_NA*','', pw2015_t2_relationship$relationship3_age_new)
pw2015_t2_relationship$relationship3_age_new <- sub('NA_*','', pw2015_t2_relationship$relationship3_age_new)
pw2015_t2_relationship$relationship3_age_new[pw2015_t2_relationship$relationship3_age_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', relationship3_age_new)) %>% separate(relationship3_age_new, c("relationship3_age_new","relationship6_age_new"), sep = "_")

pw2015_t2_relationship$relationship3_height_new <- sub('_NA*','', pw2015_t2_relationship$relationship3_height_new)
pw2015_t2_relationship$relationship3_height_new <- sub('NA_*','', pw2015_t2_relationship$relationship3_height_new)
pw2015_t2_relationship$relationship3_height_new[pw2015_t2_relationship$relationship3_height_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', relationship3_height_new)) %>% separate(relationship3_height_new, c("relationship3_height_new","relationship6_height_new"), sep = "_")

pw2015_t2_relationship$relationship3_relationship_new <- sub('_NA*','', pw2015_t2_relationship$relationship3_relationship_new)
pw2015_t2_relationship$relationship3_relationship_new <- sub('NA_*','', pw2015_t2_relationship$relationship3_relationship_new)
pw2015_t2_relationship$relationship3_relationship_new[pw2015_t2_relationship$relationship3_relationship_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', relationship3_relationship_new)) %>% separate(relationship3_relationship_new, c("relationship3_relationship_new","relationship6_relationship_new"), sep = "_")

pw2015_t2_relationship$relationship3_duration_new <- sub('_NA*','', pw2015_t2_relationship$relationship3_duration_new)
pw2015_t2_relationship$relationship3_duration_new <- sub('NA_*','', pw2015_t2_relationship$relationship3_duration_new)
pw2015_t2_relationship$relationship3_duration_new[pw2015_t2_relationship$relationship3_duration_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', relationship3_duration_new)) %>% separate(relationship3_duration_new, c("relationship3_duration_new","relationship6_duration_new"), sep = "_")
```



###Divide past relationships depending on sex

Ended relationship(s) since T1?
```{r}
#Achtung: Umgedrehte Kodierung wie bei t2_relationship!!
pw2015_t2_relationship <-  pw2015_t2_relationship %>% mutate(
  "t2_pastrelationship" = derivedFactor(
            "0" = (pastrelationship1_f == 2 & pastrelationship1_m == 2) | (pastrelationship1_f == 2 & 
                    is.na(pastrelationship1_m)) | (is.na(pastrelationship1_f) & pastrelationship1_m == 2),
            "1" = (pastrelationship1_f == 1 | pastrelationship1_m == 1), .method = "first", .default = 0))

```

Relationship 1
```{r}
pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship6_kind" = derivedVariable(
     "1" = (!is.na(pastrelationship1_kind_m) & !is.na(pastrelationship1_kind_f)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship6_kind[pw2015_t2_relationship$pastrelationship6_kind == 1] <- pw2015_t2_relationship$pastrelationship1_kind_f[pw2015_t2_relationship$pastrelationship6_kind == 1]
pw2015_t2_relationship$pastrelationship6_kind[pw2015_t2_relationship$pastrelationship6_kind == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship6_start" = derivedVariable(
     "1" = (!is.na(pastrelationship1_start_m) & !is.na(pastrelationship1_start_f)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship6_start[pw2015_t2_relationship$pastrelationship6_start == 1] <- pw2015_t2_relationship$pastrelationship1_start_f[pw2015_t2_relationship$pastrelationship6_start == 1]
pw2015_t2_relationship$pastrelationship6_start[pw2015_t2_relationship$pastrelationship6_start == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship6_end" = derivedVariable(
     "1" = (!is.na(pastrelationship1_end_m) & !is.na(pastrelationship1_end_f)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship6_end[pw2015_t2_relationship$pastrelationship6_end == 1] <- pw2015_t2_relationship$pastrelationship1_end_f[pw2015_t2_relationship$pastrelationship6_end == 1]
pw2015_t2_relationship$pastrelationship6_end[pw2015_t2_relationship$pastrelationship6_end == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship6_name" = derivedVariable(
     "1" = (!is.na(pastrelationship1_m_name) & !is.na(pastrelationship1_f_name)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship6_name[pw2015_t2_relationship$pastrelationship6_name == 1] <- pw2015_t2_relationship$pastrelationship1_f_name[pw2015_t2_relationship$pastrelationship6_name == 1]
pw2015_t2_relationship$pastrelationship6_name[pw2015_t2_relationship$pastrelationship6_name == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship6_name_new" = derivedVariable(
     "1" = (!is.na(pastrelationship1_name_m_new) & !is.na(pastrelationship1_name_f_new)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship6_name_new[pw2015_t2_relationship$pastrelationship6_name_new == 1] <- pw2015_t2_relationship$pastrelationship1_name_f_new[pw2015_t2_relationship$pastrelationship6_name_new == 1]
pw2015_t2_relationship$pastrelationship6_name_new[pw2015_t2_relationship$pastrelationship6_name_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship6_age_new" = derivedVariable(
     "1" = (!is.na(pastrelationship1_age_m_new) & !is.na(pastrelationship1_age_f_new)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship6_age_new[pw2015_t2_relationship$pastrelationship6_age_new == 1] <- pw2015_t2_relationship$pastrelationship1_age_f_new[pw2015_t2_relationship$pastrelationship6_age_new == 1]
pw2015_t2_relationship$pastrelationship6_age_new[pw2015_t2_relationship$pastrelationship6_age_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship6_height_new" = derivedVariable(
     "1" = (!is.na(pastrelationship1_height_m_new) & !is.na(pastrelationship1_height_f_new)),
     .method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship6_height_new[pw2015_t2_relationship$pastrelationship6_height_new == 1] <- pw2015_t2_relationship$pastrelationship1_height_f_new[pw2015_t2_relationship$pastrelationship6_height_new == 1]
pw2015_t2_relationship$pastrelationship6_height_new[pw2015_t2_relationship$pastrelationship6_height_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship6_duration_new" = derivedVariable(
     "1" = (!is.na(pastrelationship1_duration_m_new) & !is.na(pastrelationship1_duration_f_new)),
     .method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship6_duration_new[pw2015_t2_relationship$pastrelationship6_duration_new == 1] <- pw2015_t2_relationship$pastrelationship1_duration_f_new[pw2015_t2_relationship$pastrelationship6_duration_new == 1]
pw2015_t2_relationship$pastrelationship6_duration_new[pw2015_t2_relationship$pastrelationship6_duration_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship6_relationship_new" = derivedVariable(
     "1" = (!is.na(pastrelationship1_relationship_m_new) & !is.na(pastrelationship1_relationship_f_new)),
     .method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship6_relationship_new[pw2015_t2_relationship$pastrelationship6_relationship_new == 1] <- pw2015_t2_relationship$pastrelationship1_relationship_f_new[pw2015_t2_relationship$pastrelationship6_relationship_new == 1]
pw2015_t2_relationship$pastrelationship6_relationship_new[pw2015_t2_relationship$pastrelationship6_relationship_new == 0] <- NA
```

Relationship 2
```{r}
pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship7_kind" = derivedVariable(
     "1" = (!is.na(pastrelationship2_kind_m) & !is.na(pastrelationship2_kind_f)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship7_kind[pw2015_t2_relationship$pastrelationship7_kind == 1] <- pw2015_t2_relationship$pastrelationship2_kind_f[pw2015_t2_relationship$pastrelationship7_kind == 1]
pw2015_t2_relationship$pastrelationship7_kind[pw2015_t2_relationship$pastrelationship7_kind == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship7_start" = derivedVariable(
     "1" = (!is.na(pastrelationship2_start_m) & !is.na(pastrelationship2_start_f)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship7_start[pw2015_t2_relationship$pastrelationship7_start == 1] <- pw2015_t2_relationship$pastrelationship2_start_f[pw2015_t2_relationship$pastrelationship7_start == 1]
pw2015_t2_relationship$pastrelationship7_start[pw2015_t2_relationship$pastrelationship7_start == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship7_end" = derivedVariable(
     "1" = (!is.na(pastrelationship2_end_m) & !is.na(pastrelationship2_end_f)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship7_end[pw2015_t2_relationship$pastrelationship7_end == 1] <- pw2015_t2_relationship$pastrelationship2_end_f[pw2015_t2_relationship$pastrelationship7_end == 1]
pw2015_t2_relationship$pastrelationship7_end[pw2015_t2_relationship$pastrelationship7_end == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship7_name" = derivedVariable(
     "1" = (!is.na(pastrelationship2_m_name) & !is.na(pastrelationship2_f_name)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship7_name[pw2015_t2_relationship$pastrelationship7_name == 1] <- pw2015_t2_relationship$pastrelationship2_f_name[pw2015_t2_relationship$pastrelationship7_name == 1]
pw2015_t2_relationship$pastrelationship7_name[pw2015_t2_relationship$pastrelationship7_name == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship7_name_new" = derivedVariable(
     "1" = (!is.na(pastrelationship2_name_m_new) & !is.na(pastrelationship2_name_f_new)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship7_name_new[pw2015_t2_relationship$pastrelationship7_name_new == 1] <- pw2015_t2_relationship$pastrelationship2_name_f_new[pw2015_t2_relationship$pastrelationship7_name_new == 1]
pw2015_t2_relationship$pastrelationship7_name_new[pw2015_t2_relationship$pastrelationship7_name_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship7_age_new" = derivedVariable(
     "1" = (!is.na(pastrelationship2_age_m_new) & !is.na(pastrelationship2_age_f_new)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship7_age_new[pw2015_t2_relationship$pastrelationship7_age_new == 1] <- pw2015_t2_relationship$pastrelationship2_age_f_new[pw2015_t2_relationship$pastrelationship7_age_new == 1]
pw2015_t2_relationship$pastrelationship7_age_new[pw2015_t2_relationship$pastrelationship7_age_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship7_height_new" = derivedVariable(
     "1" = (!is.na(pastrelationship2_height_m_new) & !is.na(pastrelationship2_height_f_new)),
     .method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship7_height_new[pw2015_t2_relationship$pastrelationship7_height_new == 1] <- pw2015_t2_relationship$pastrelationship2_height_f_new[pw2015_t2_relationship$pastrelationship7_height_new == 1]
pw2015_t2_relationship$pastrelationship7_height_new[pw2015_t2_relationship$pastrelationship7_height_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship7_duration_new" = derivedVariable(
     "1" = (!is.na(pastrelationship2_duration_m_new) & !is.na(pastrelationship2_duration_f_new)),
     .method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship7_duration_new[pw2015_t2_relationship$pastrelationship7_duration_new == 1] <- pw2015_t2_relationship$pastrelationship2_duration_f_new[pw2015_t2_relationship$pastrelationship7_duration_new == 1]
pw2015_t2_relationship$pastrelationship7_duration_new[pw2015_t2_relationship$pastrelationship7_duration_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship7_relationship_new" = derivedVariable(
     "1" = (!is.na(pastrelationship2_relationship_m_new) & !is.na(pastrelationship2_relationship_f_new)),
     .method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship7_relationship_new[pw2015_t2_relationship$pastrelationship7_relationship_new == 1] <- pw2015_t2_relationship$pastrelationship2_relationship_f_new[pw2015_t2_relationship$pastrelationship7_relationship_new == 1]
pw2015_t2_relationship$pastrelationship7_relationship_new[pw2015_t2_relationship$pastrelationship7_relationship_new == 0] <- NA
```

Relationship 3
```{r}
pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship8_kind" = derivedVariable(
     "1" = (!is.na(pastrelationship3_kind_m) & !is.na(pastrelationship3_kind_f)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship8_kind[pw2015_t2_relationship$pastrelationship8_kind == 1] <- pw2015_t2_relationship$pastrelationship3_kind_f[pw2015_t2_relationship$pastrelationship8_kind == 1]
pw2015_t2_relationship$pastrelationship8_kind[pw2015_t2_relationship$pastrelationship8_kind == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship8_start" = derivedVariable(
     "1" = (!is.na(pastrelationship3_start_m) & !is.na(pastrelationship3_start_f)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship8_start[pw2015_t2_relationship$pastrelationship8_start == 1] <- pw2015_t2_relationship$pastrelationship3_start_f[pw2015_t2_relationship$pastrelationship8_start == 1]
pw2015_t2_relationship$pastrelationship8_start[pw2015_t2_relationship$pastrelationship8_start == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship8_end" = derivedVariable(
     "1" = (!is.na(pastrelationship3_end_m) & !is.na(pastrelationship3_end_f)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship8_end[pw2015_t2_relationship$pastrelationship8_end == 1] <- pw2015_t2_relationship$pastrelationship3_end_f[pw2015_t2_relationship$pastrelationship8_end == 1]
pw2015_t2_relationship$pastrelationship8_end[pw2015_t2_relationship$pastrelationship8_end == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship8_name" = derivedVariable(
     "1" = (!is.na(pastrelationship3_m_name) & !is.na(pastrelationship3_f_name)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship8_name[pw2015_t2_relationship$pastrelationship8_name == 1] <- pw2015_t2_relationship$pastrelationship3_f_name[pw2015_t2_relationship$pastrelationship8_name == 1]
pw2015_t2_relationship$pastrelationship8_name[pw2015_t2_relationship$pastrelationship8_name == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship8_name_new" = derivedVariable(
     "1" = (!is.na(pastrelationship3_name_m_new) & !is.na(pastrelationship3_name_f_new)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship8_name_new[pw2015_t2_relationship$pastrelationship8_name_new == 1] <- pw2015_t2_relationship$pastrelationship3_name_f_new[pw2015_t2_relationship$pastrelationship8_name_new == 1]
pw2015_t2_relationship$pastrelationship8_name_new[pw2015_t2_relationship$pastrelationship8_name_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship8_age_new" = derivedVariable(
     "1" = (!is.na(pastrelationship3_age_m_new) & !is.na(pastrelationship3_age_f_new)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship8_age_new[pw2015_t2_relationship$pastrelationship8_age_new == 1] <- pw2015_t2_relationship$pastrelationship3_age_f_new[pw2015_t2_relationship$pastrelationship8_age_new == 1]
pw2015_t2_relationship$pastrelationship8_age_new[pw2015_t2_relationship$pastrelationship8_age_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship8_height_new" = derivedVariable(
     "1" = (!is.na(pastrelationship3_height_m_new) & !is.na(pastrelationship3_height_f_new)),
     .method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship8_height_new[pw2015_t2_relationship$pastrelationship8_height_new == 1] <- pw2015_t2_relationship$pastrelationship3_height_f_new[pw2015_t2_relationship$pastrelationship8_height_new == 1]
pw2015_t2_relationship$pastrelationship8_height_new[pw2015_t2_relationship$pastrelationship8_height_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship8_duration_new" = derivedVariable(
     "1" = (!is.na(pastrelationship3_duration_m_new) & !is.na(pastrelationship3_duration_f_new)),
     .method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship8_duration_new[pw2015_t2_relationship$pastrelationship8_duration_new == 1] <- pw2015_t2_relationship$pastrelationship3_duration_f_new[pw2015_t2_relationship$pastrelationship8_duration_new == 1]
pw2015_t2_relationship$pastrelationship8_duration_new[pw2015_t2_relationship$pastrelationship8_duration_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship8_relationship_new" = derivedVariable(
     "1" = (!is.na(pastrelationship3_relationship_m_new) & !is.na(pastrelationship3_relationship_f_new)),
     .method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship8_relationship_new[pw2015_t2_relationship$pastrelationship8_relationship_new == 1] <- pw2015_t2_relationship$pastrelationship3_relationship_f_new[pw2015_t2_relationship$pastrelationship8_relationship_new == 1]
pw2015_t2_relationship$pastrelationship8_relationship_new[pw2015_t2_relationship$pastrelationship8_relationship_new == 0] <- NA
```

Relationship 4
```{r}
pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship9_kind" = derivedVariable(
     "1" = (!is.na(pastrelationship4_kind_m) & !is.na(pastrelationship4_kind_f)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship9_kind[pw2015_t2_relationship$pastrelationship9_kind == 1] <- pw2015_t2_relationship$pastrelationship4_kind_f[pw2015_t2_relationship$pastrelationship9_kind == 1]
pw2015_t2_relationship$pastrelationship9_kind[pw2015_t2_relationship$pastrelationship9_kind == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship9_start" = derivedVariable(
     "1" = (!is.na(pastrelationship4_start_m) & !is.na(pastrelationship4_start_f)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship9_start[pw2015_t2_relationship$pastrelationship9_start == 1] <- pw2015_t2_relationship$pastrelationship4_start_f[pw2015_t2_relationship$pastrelationship9_start == 1]
pw2015_t2_relationship$pastrelationship9_start[pw2015_t2_relationship$pastrelationship9_start == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship9_end" = derivedVariable(
     "1" = (!is.na(pastrelationship4_end_m) & !is.na(pastrelationship4_end_f)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship9_end[pw2015_t2_relationship$pastrelationship9_end == 1] <- pw2015_t2_relationship$pastrelationship4_end_f[pw2015_t2_relationship$pastrelationship9_end == 1]
pw2015_t2_relationship$pastrelationship9_end[pw2015_t2_relationship$pastrelationship9_end == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship9_name" = derivedVariable(
     "1" = (!is.na(pastrelationship4_m_name) & !is.na(pastrelationship4_f_name)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship9_name[pw2015_t2_relationship$pastrelationship9_name == 1] <- pw2015_t2_relationship$pastrelationship4_f_name[pw2015_t2_relationship$pastrelationship9_name == 1]
pw2015_t2_relationship$pastrelationship9_name[pw2015_t2_relationship$pastrelationship9_name == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship9_name_new" = derivedVariable(
     "1" = (!is.na(pastrelationship4_name_m_new) & !is.na(pastrelationship4_name_f_new)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship9_name_new[pw2015_t2_relationship$pastrelationship9_name_new == 1] <- pw2015_t2_relationship$pastrelationship4_name_f_new[pw2015_t2_relationship$pastrelationship9_name_new == 1]
pw2015_t2_relationship$pastrelationship9_name_new[pw2015_t2_relationship$pastrelationship9_name_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship9_age_new" = derivedVariable(
     "1" = (!is.na(pastrelationship4_age_m_new) & !is.na(pastrelationship4_age_f_new)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship9_age_new[pw2015_t2_relationship$pastrelationship9_age_new == 1] <- pw2015_t2_relationship$pastrelationship4_age_f_new[pw2015_t2_relationship$pastrelationship9_age_new == 1]
pw2015_t2_relationship$pastrelationship9_age_new[pw2015_t2_relationship$pastrelationship9_age_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship9_height_new" = derivedVariable(
     "1" = (!is.na(pastrelationship4_height_m_new) & !is.na(pastrelationship4_height_f_new)),
     .method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship9_height_new[pw2015_t2_relationship$pastrelationship9_height_new == 1] <- pw2015_t2_relationship$pastrelationship4_height_f_new[pw2015_t2_relationship$pastrelationship9_height_new == 1]
pw2015_t2_relationship$pastrelationship9_height_new[pw2015_t2_relationship$pastrelationship9_height_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship9_duration_new" = derivedVariable(
     "1" = (!is.na(pastrelationship4_duration_m_new) & !is.na(pastrelationship4_duration_f_new)),
     .method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship9_duration_new[pw2015_t2_relationship$pastrelationship9_duration_new == 1] <- pw2015_t2_relationship$pastrelationship4_duration_f_new[pw2015_t2_relationship$pastrelationship9_duration_new == 1]
pw2015_t2_relationship$pastrelationship9_duration_new[pw2015_t2_relationship$pastrelationship9_duration_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship9_relationship_new" = derivedVariable(
     "1" = (!is.na(pastrelationship4_relationship_m_new) & !is.na(pastrelationship4_relationship_f_new)),
     .method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship9_relationship_new[pw2015_t2_relationship$pastrelationship9_relationship_new == 1] <- pw2015_t2_relationship$pastrelationship4_relationship_f_new[pw2015_t2_relationship$pastrelationship9_relationship_new == 1]
pw2015_t2_relationship$pastrelationship9_relationship_new[pw2015_t2_relationship$pastrelationship9_relationship_new == 0] <- NA
```

Relationship 5
```{r}
pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship10_kind" = derivedVariable(
     "1" = (!is.na(pastrelationship5_kind_m) & !is.na(pastrelationship5_kind_f)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship10_kind[pw2015_t2_relationship$pastrelationship10_kind == 1] <- pw2015_t2_relationship$pastrelationship5_kind_f[pw2015_t2_relationship$pastrelationship10_kind == 1]
pw2015_t2_relationship$pastrelationship10_kind[pw2015_t2_relationship$pastrelationship10_kind == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship10_start" = derivedVariable(
     "1" = (!is.na(pastrelationship5_start_m) & !is.na(pastrelationship5_start_f)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship10_start[pw2015_t2_relationship$pastrelationship10_start == 1] <- pw2015_t2_relationship$pastrelationship5_start_f[pw2015_t2_relationship$pastrelationship10_start == 1]
pw2015_t2_relationship$pastrelationship10_start[pw2015_t2_relationship$pastrelationship10_start == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship10_end" = derivedVariable(
     "1" = (!is.na(pastrelationship5_end_m) & !is.na(pastrelationship5_end_f)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship10_end[pw2015_t2_relationship$pastrelationship10_end == 1] <- pw2015_t2_relationship$pastrelationship5_end_f[pw2015_t2_relationship$pastrelationship10_end == 1]
pw2015_t2_relationship$pastrelationship10_end[pw2015_t2_relationship$pastrelationship10_end == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship10_name" = derivedVariable(
     "1" = (!is.na(pastrelationship5_m_name) & !is.na(pastrelationship5_f_name)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship10_name[pw2015_t2_relationship$pastrelationship10_name == 1] <- pw2015_t2_relationship$pastrelationship5_f_name[pw2015_t2_relationship$pastrelationship10_name == 1]
pw2015_t2_relationship$pastrelationship10_name[pw2015_t2_relationship$pastrelationship10_name == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship10_name_new" = derivedVariable(
     "1" = (!is.na(pastrelationship5_name_m_new) & !is.na(pastrelationship5_name_f_new)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship10_name_new[pw2015_t2_relationship$pastrelationship10_name_new == 1] <- pw2015_t2_relationship$pastrelationship5_name_f_new[pw2015_t2_relationship$pastrelationship10_name_new == 1]
pw2015_t2_relationship$pastrelationship10_name_new[pw2015_t2_relationship$pastrelationship10_name_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship10_age_new" = derivedVariable(
     "1" = (!is.na(pastrelationship5_age_m_new) & !is.na(pastrelationship5_age_f_new)),.method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship10_age_new[pw2015_t2_relationship$pastrelationship10_age_new == 1] <- pw2015_t2_relationship$pastrelationship5_age_f_new[pw2015_t2_relationship$pastrelationship10_age_new == 1]
pw2015_t2_relationship$pastrelationship10_age_new[pw2015_t2_relationship$pastrelationship10_age_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship10_height_new" = derivedVariable(
     "1" = (!is.na(pastrelationship5_height_m_new) & !is.na(pastrelationship5_height_f_new)),
     .method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship10_height_new[pw2015_t2_relationship$pastrelationship10_height_new == 1] <- pw2015_t2_relationship$pastrelationship5_height_f_new[pw2015_t2_relationship$pastrelationship10_height_new == 1]
pw2015_t2_relationship$pastrelationship10_height_new[pw2015_t2_relationship$pastrelationship10_height_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship10_duration_new" = derivedVariable(
     "1" = (!is.na(pastrelationship5_duration_m_new) & !is.na(pastrelationship5_duration_f_new)),
     .method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship10_duration_new[pw2015_t2_relationship$pastrelationship10_duration_new == 1] <- pw2015_t2_relationship$pastrelationship5_duration_f_new[pw2015_t2_relationship$pastrelationship10_duration_new == 1]
pw2015_t2_relationship$pastrelationship10_duration_new[pw2015_t2_relationship$pastrelationship10_duration_new == 0] <- NA

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "pastrelationship10_relationship_new" = derivedVariable(
     "1" = (!is.na(pastrelationship5_relationship_m_new) & !is.na(pastrelationship5_relationship_f_new)),
     .method = "first", .default = 0))
pw2015_t2_relationship$pastrelationship10_relationship_new[pw2015_t2_relationship$pastrelationship10_relationship_new == 1] <- pw2015_t2_relationship$pastrelationship5_relationship_f_new[pw2015_t2_relationship$pastrelationship10_relationship_new == 1]
pw2015_t2_relationship$pastrelationship10_relationship_new[pw2015_t2_relationship$pastrelationship10_relationship_new == 0] <- NA
```




Unite "pastrelationship"-variables
```{r}
#Relationship 1
pw2015_t2_relationship <- pw2015_t2_relationship %>% unite(pastrelationship1_name, pastrelationship1_m_name, pastrelationship1_f_name) %>%
      unite(pastrelationship1_name_new, pastrelationship1_name_m_new, pastrelationship1_name_f_new) %>% 
      unite(pastrelationship1_age_new, pastrelationship1_age_m_new, pastrelationship1_age_f_new) %>% 
      unite(pastrelationship1_height_new, pastrelationship1_height_m_new, pastrelationship1_height_f_new) %>% 
      unite(pastrelationship1_relationship_new, pastrelationship1_relationship_m_new, pastrelationship1_relationship_f_new) %>% 
      unite(pastrelationship1_duration_new, pastrelationship1_duration_m_new, pastrelationship1_duration_f_new) %>%
      unite(pastrelationship1_kind, pastrelationship1_kind_m, pastrelationship1_kind_f) %>%
      unite(pastrelationship1_start, pastrelationship1_start_m, pastrelationship1_start_f) %>%
      unite(pastrelationship1_end, pastrelationship1_end_m, pastrelationship1_end_f)

#Relationship 2
pw2015_t2_relationship <- pw2015_t2_relationship %>% unite(pastrelationship2_name, pastrelationship2_m_name, pastrelationship2_f_name) %>%
      unite(pastrelationship2_name_new, pastrelationship2_name_m_new, pastrelationship2_name_f_new) %>% 
      unite(pastrelationship2_age_new, pastrelationship2_age_m_new, pastrelationship2_age_f_new) %>% 
      unite(pastrelationship2_height_new, pastrelationship2_height_m_new, pastrelationship2_height_f_new) %>% 
      unite(pastrelationship2_relationship_new, pastrelationship2_relationship_m_new, pastrelationship2_relationship_f_new) %>% 
      unite(pastrelationship2_duration_new, pastrelationship2_duration_m_new, pastrelationship2_duration_f_new) %>%
      unite(pastrelationship2_kind, pastrelationship2_kind_m, pastrelationship2_kind_f) %>%
      unite(pastrelationship2_start, pastrelationship2_start_m, pastrelationship2_start_f) %>%
      unite(pastrelationship2_end, pastrelationship2_end_m, pastrelationship2_end_f)

#Relationship 3
pw2015_t2_relationship <- pw2015_t2_relationship %>% unite(pastrelationship3_name, pastrelationship3_m_name, pastrelationship3_f_name) %>%
      unite(pastrelationship3_name_new, pastrelationship3_name_m_new, pastrelationship3_name_f_new) %>% 
      unite(pastrelationship3_age_new, pastrelationship3_age_m_new, pastrelationship3_age_f_new) %>% 
      unite(pastrelationship3_height_new, pastrelationship3_height_m_new, pastrelationship3_height_f_new) %>% 
      unite(pastrelationship3_relationship_new, pastrelationship3_relationship_m_new, pastrelationship3_relationship_f_new) %>% 
      unite(pastrelationship3_duration_new, pastrelationship3_duration_m_new, pastrelationship3_duration_f_new) %>%
      unite(pastrelationship3_kind, pastrelationship3_kind_m, pastrelationship3_kind_f) %>%
      unite(pastrelationship3_start, pastrelationship3_start_m, pastrelationship3_start_f) %>%
      unite(pastrelationship3_end, pastrelationship3_end_m, pastrelationship3_end_f)

#Relationship 4
pw2015_t2_relationship <- pw2015_t2_relationship %>% unite(pastrelationship4_name, pastrelationship4_m_name, pastrelationship4_f_name) %>%
      unite(pastrelationship4_name_new, pastrelationship4_name_m_new, pastrelationship4_name_f_new) %>% 
      unite(pastrelationship4_age_new, pastrelationship4_age_m_new, pastrelationship4_age_f_new) %>% 
      unite(pastrelationship4_height_new, pastrelationship4_height_m_new, pastrelationship4_height_f_new) %>% 
      unite(pastrelationship4_relationship_new, pastrelationship4_relationship_m_new, pastrelationship4_relationship_f_new) %>% 
      unite(pastrelationship4_duration_new, pastrelationship4_duration_m_new, pastrelationship4_duration_f_new) %>%
      unite(pastrelationship4_kind, pastrelationship4_kind_m, pastrelationship4_kind_f) %>%
      unite(pastrelationship4_start, pastrelationship4_start_m, pastrelationship4_start_f) %>%
      unite(pastrelationship4_end, pastrelationship4_end_m, pastrelationship4_end_f)

#Relationship 5
pw2015_t2_relationship <- pw2015_t2_relationship %>% unite(pastrelationship5_name, pastrelationship5_m_name, pastrelationship5_f_name) %>%
      unite(pastrelationship5_name_new, pastrelationship5_name_m_new, pastrelationship5_name_f_new) %>% 
      unite(pastrelationship5_age_new, pastrelationship5_age_m_new, pastrelationship5_age_f_new) %>% 
      unite(pastrelationship5_height_new, pastrelationship5_height_m_new, pastrelationship5_height_f_new) %>% 
      unite(pastrelationship5_relationship_new, pastrelationship5_relationship_m_new, pastrelationship5_relationship_f_new) %>% 
      unite(pastrelationship5_duration_new, pastrelationship5_duration_m_new, pastrelationship5_duration_f_new) %>%
      unite(pastrelationship5_kind, pastrelationship5_kind_m, pastrelationship5_kind_f) %>%
      unite(pastrelationship5_start, pastrelationship5_start_m, pastrelationship5_start_f) %>%
      unite(pastrelationship5_end, pastrelationship5_end_m, pastrelationship5_end_f)
```

Divide variables, necessary when VPs had both male and female partners

Past relationship 1
```{r}
pw2015_t2_relationship$pastrelationship1_kind <- sub('_NA*','', pw2015_t2_relationship$pastrelationship1_kind)
pw2015_t2_relationship$pastrelationship1_kind <- sub('NA_*','', pw2015_t2_relationship$pastrelationship1_kind)
pw2015_t2_relationship$pastrelationship1_kind[pw2015_t2_relationship$pastrelationship1_kind==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship1_kind)) %>% separate(pastrelationship1_kind, c("pastrelationship1_kind","pastrelationship6_kind"), sep = "_")

pw2015_t2_relationship$pastrelationship1_start <- sub('_NA*','', pw2015_t2_relationship$pastrelationship1_start)
pw2015_t2_relationship$pastrelationship1_start <- sub('NA_*','', pw2015_t2_relationship$pastrelationship1_start)
pw2015_t2_relationship$pastrelationship1_start[pw2015_t2_relationship$pastrelationship1_start==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship1_start)) %>% separate(pastrelationship1_start, c("pastrelationship1_start","pastrelationship6_start"), sep = "_")

pw2015_t2_relationship$pastrelationship1_end <- sub('_NA*','', pw2015_t2_relationship$pastrelationship1_end)
pw2015_t2_relationship$pastrelationship1_end <- sub('NA_*','', pw2015_t2_relationship$pastrelationship1_end)
pw2015_t2_relationship$pastrelationship1_end[pw2015_t2_relationship$pastrelationship1_end==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship1_end)) %>% separate(pastrelationship1_end, c("pastrelationship1_end","pastrelationship6_end"), sep = "_")

pw2015_t2_relationship$pastrelationship1_name <- sub('_NA*','', pw2015_t2_relationship$pastrelationship1_name)
pw2015_t2_relationship$pastrelationship1_name <- sub('NA_*','', pw2015_t2_relationship$pastrelationship1_name)
pw2015_t2_relationship$pastrelationship1_name[pw2015_t2_relationship$pastrelationship1_name==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship1_name)) %>% separate(pastrelationship1_name, c("pastrelationship1_name","pastrelationship6_name"), sep = "_")

pw2015_t2_relationship$pastrelationship1_name_new <- sub('_NA*','', pw2015_t2_relationship$pastrelationship1_name_new)
pw2015_t2_relationship$pastrelationship1_name_new <- sub('NA_*','', pw2015_t2_relationship$pastrelationship1_name_new)
pw2015_t2_relationship$pastrelationship1_name_new[pw2015_t2_relationship$pastrelationship1_name_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship1_name_new)) %>% separate(pastrelationship1_name_new, c("pastrelationship1_name_new","pastrelationship6_name_new"), sep = "_")

pw2015_t2_relationship$pastrelationship1_age_new <- sub('_NA*','', pw2015_t2_relationship$pastrelationship1_age_new)
pw2015_t2_relationship$pastrelationship1_age_new <- sub('NA_*','', pw2015_t2_relationship$pastrelationship1_age_new)
pw2015_t2_relationship$pastrelationship1_age_new[pw2015_t2_relationship$pastrelationship1_age_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship1_age_new)) %>% separate(pastrelationship1_age_new, c("pastrelationship1_age_new","pastrelationship6_age_new"), sep = "_")

pw2015_t2_relationship$pastrelationship1_height_new <- sub('_NA*','', pw2015_t2_relationship$pastrelationship1_height_new)
pw2015_t2_relationship$pastrelationship1_height_new <- sub('NA_*','', pw2015_t2_relationship$pastrelationship1_height_new)
pw2015_t2_relationship$pastrelationship1_height_new[pw2015_t2_relationship$pastrelationship1_height_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship1_height_new)) %>% separate(pastrelationship1_height_new, c("pastrelationship1_height_new","pastrelationship6_height_new"), sep = "_")

pw2015_t2_relationship$pastrelationship1_relationship_new <- sub('_NA*','', pw2015_t2_relationship$pastrelationship1_relationship_new)
pw2015_t2_relationship$pastrelationship1_relationship_new <- sub('NA_*','', pw2015_t2_relationship$pastrelationship1_relationship_new)
pw2015_t2_relationship$pastrelationship1_relationship_new[pw2015_t2_relationship$pastrelationship1_relationship_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship1_relationship_new)) %>% separate(pastrelationship1_relationship_new, c("pastrelationship1_relationship_new","pastrelationship6_relationship_new"), sep = "_")

pw2015_t2_relationship$pastrelationship1_duration_new <- sub('_NA*','', pw2015_t2_relationship$pastrelationship1_duration_new)
pw2015_t2_relationship$pastrelationship1_duration_new <- sub('NA_*','', pw2015_t2_relationship$pastrelationship1_duration_new)
pw2015_t2_relationship$pastrelationship1_duration_new[pw2015_t2_relationship$pastrelationship1_duration_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship1_duration_new)) %>% separate(pastrelationship1_duration_new, c("pastrelationship1_duration_new","pastrelationship6_duration_new"), sep = "_")
```

Past relationship 2
```{r}
pw2015_t2_relationship$pastrelationship2_kind <- sub('_NA*','', pw2015_t2_relationship$pastrelationship2_kind)
pw2015_t2_relationship$pastrelationship2_kind <- sub('NA_*','', pw2015_t2_relationship$pastrelationship2_kind)
pw2015_t2_relationship$pastrelationship2_kind[pw2015_t2_relationship$pastrelationship2_kind==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship2_kind)) %>% separate(pastrelationship2_kind, c("pastrelationship2_kind","pastrelationship7_kind"), sep = "_")

pw2015_t2_relationship$pastrelationship2_start <- sub('_NA*','', pw2015_t2_relationship$pastrelationship2_start)
pw2015_t2_relationship$pastrelationship2_start <- sub('NA_*','', pw2015_t2_relationship$pastrelationship2_start)
pw2015_t2_relationship$pastrelationship2_start[pw2015_t2_relationship$pastrelationship2_start==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship2_start)) %>% separate(pastrelationship2_start, c("pastrelationship2_start","pastrelationship7_start"), sep = "_")

pw2015_t2_relationship$pastrelationship2_end <- sub('_NA*','', pw2015_t2_relationship$pastrelationship2_end)
pw2015_t2_relationship$pastrelationship2_end <- sub('NA_*','', pw2015_t2_relationship$pastrelationship2_end)
pw2015_t2_relationship$pastrelationship2_end[pw2015_t2_relationship$pastrelationship2_end==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship2_end)) %>% separate(pastrelationship2_end, c("pastrelationship2_end","pastrelationship7_end"), sep = "_")

pw2015_t2_relationship$pastrelationship2_name <- sub('_NA*','', pw2015_t2_relationship$pastrelationship2_name)
pw2015_t2_relationship$pastrelationship2_name <- sub('NA_*','', pw2015_t2_relationship$pastrelationship2_name)
pw2015_t2_relationship$pastrelationship2_name[pw2015_t2_relationship$pastrelationship2_name==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship2_name)) %>% separate(pastrelationship2_name, c("pastrelationship2_name","pastrelationship7_name"), sep = "_")

pw2015_t2_relationship$pastrelationship2_name_new <- sub('_NA*','', pw2015_t2_relationship$pastrelationship2_name_new)
pw2015_t2_relationship$pastrelationship2_name_new <- sub('NA_*','', pw2015_t2_relationship$pastrelationship2_name_new)
pw2015_t2_relationship$pastrelationship2_name_new[pw2015_t2_relationship$pastrelationship2_name_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship2_name_new)) %>% separate(pastrelationship2_name_new, c("pastrelationship2_name_new","pastrelationship7_name_new"), sep = "_")

pw2015_t2_relationship$pastrelationship2_age_new <- sub('_NA*','', pw2015_t2_relationship$pastrelationship2_age_new)
pw2015_t2_relationship$pastrelationship2_age_new <- sub('NA_*','', pw2015_t2_relationship$pastrelationship2_age_new)
pw2015_t2_relationship$pastrelationship2_age_new[pw2015_t2_relationship$pastrelationship2_age_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship2_age_new)) %>% separate(pastrelationship2_age_new, c("pastrelationship2_age_new","pastrelationship7_age_new"), sep = "_")

pw2015_t2_relationship$pastrelationship2_height_new <- sub('_NA*','', pw2015_t2_relationship$pastrelationship2_height_new)
pw2015_t2_relationship$pastrelationship2_height_new <- sub('NA_*','', pw2015_t2_relationship$pastrelationship2_height_new)
pw2015_t2_relationship$pastrelationship2_height_new[pw2015_t2_relationship$pastrelationship2_height_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship2_height_new)) %>% separate(pastrelationship2_height_new, c("pastrelationship2_height_new","pastrelationship7_height_new"), sep = "_")

pw2015_t2_relationship$pastrelationship2_relationship_new <- sub('_NA*','', pw2015_t2_relationship$pastrelationship2_relationship_new)
pw2015_t2_relationship$pastrelationship2_relationship_new <- sub('NA_*','', pw2015_t2_relationship$pastrelationship2_relationship_new)
pw2015_t2_relationship$pastrelationship2_relationship_new[pw2015_t2_relationship$pastrelationship2_relationship_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship2_relationship_new)) %>% separate(pastrelationship2_relationship_new, c("pastrelationship2_relationship_new","pastrelationship7_relationship_new"), sep = "_")

pw2015_t2_relationship$pastrelationship2_duration_new <- sub('_NA*','', pw2015_t2_relationship$pastrelationship2_duration_new)
pw2015_t2_relationship$pastrelationship2_duration_new <- sub('NA_*','', pw2015_t2_relationship$pastrelationship2_duration_new)
pw2015_t2_relationship$pastrelationship2_duration_new[pw2015_t2_relationship$pastrelationship2_duration_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship2_duration_new)) %>% separate(pastrelationship2_duration_new, c("pastrelationship2_duration_new","pastrelationship7_duration_new"), sep = "_")
```

Past relationship 3
```{r}
pw2015_t2_relationship$pastrelationship3_kind <- sub('_NA*','', pw2015_t2_relationship$pastrelationship3_kind)
pw2015_t2_relationship$pastrelationship3_kind <- sub('NA_*','', pw2015_t2_relationship$pastrelationship3_kind)
pw2015_t2_relationship$pastrelationship3_kind[pw2015_t2_relationship$pastrelationship3_kind==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship3_kind)) %>% separate(pastrelationship3_kind, c("pastrelationship3_kind","pastrelationship8_kind"), sep = "_")

pw2015_t2_relationship$pastrelationship3_start <- sub('_NA*','', pw2015_t2_relationship$pastrelationship3_start)
pw2015_t2_relationship$pastrelationship3_start <- sub('NA_*','', pw2015_t2_relationship$pastrelationship3_start)
pw2015_t2_relationship$pastrelationship3_start[pw2015_t2_relationship$pastrelationship3_start==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship3_start)) %>% separate(pastrelationship3_start, c("pastrelationship3_start","pastrelationship8_start"), sep = "_")

pw2015_t2_relationship$pastrelationship3_end <- sub('_NA*','', pw2015_t2_relationship$pastrelationship3_end)
pw2015_t2_relationship$pastrelationship3_end <- sub('NA_*','', pw2015_t2_relationship$pastrelationship3_end)
pw2015_t2_relationship$pastrelationship3_end[pw2015_t2_relationship$pastrelationship3_end==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship3_end)) %>% separate(pastrelationship3_end, c("pastrelationship3_end","pastrelationship8_end"), sep = "_")

pw2015_t2_relationship$pastrelationship3_name <- sub('_NA*','', pw2015_t2_relationship$pastrelationship3_name)
pw2015_t2_relationship$pastrelationship3_name <- sub('NA_*','', pw2015_t2_relationship$pastrelationship3_name)
pw2015_t2_relationship$pastrelationship3_name[pw2015_t2_relationship$pastrelationship3_name==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship3_name)) %>% separate(pastrelationship3_name, c("pastrelationship3_name","pastrelationship8_name"), sep = "_")

pw2015_t2_relationship$pastrelationship3_name_new <- sub('_NA*','', pw2015_t2_relationship$pastrelationship3_name_new)
pw2015_t2_relationship$pastrelationship3_name_new <- sub('NA_*','', pw2015_t2_relationship$pastrelationship3_name_new)
pw2015_t2_relationship$pastrelationship3_name_new[pw2015_t2_relationship$pastrelationship3_name_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship3_name_new)) %>% separate(pastrelationship3_name_new, c("pastrelationship3_name_new","pastrelationship8_name_new"), sep = "_")

pw2015_t2_relationship$pastrelationship3_age_new <- sub('_NA*','', pw2015_t2_relationship$pastrelationship3_age_new)
pw2015_t2_relationship$pastrelationship3_age_new <- sub('NA_*','', pw2015_t2_relationship$pastrelationship3_age_new)
pw2015_t2_relationship$pastrelationship3_age_new[pw2015_t2_relationship$pastrelationship3_age_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship3_age_new)) %>% separate(pastrelationship3_age_new, c("pastrelationship3_age_new","pastrelationship8_age_new"), sep = "_")

pw2015_t2_relationship$pastrelationship3_height_new <- sub('_NA*','', pw2015_t2_relationship$pastrelationship3_height_new)
pw2015_t2_relationship$pastrelationship3_height_new <- sub('NA_*','', pw2015_t2_relationship$pastrelationship3_height_new)
pw2015_t2_relationship$pastrelationship3_height_new[pw2015_t2_relationship$pastrelationship3_height_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship3_height_new)) %>% separate(pastrelationship3_height_new, c("pastrelationship3_height_new","pastrelationship8_height_new"), sep = "_")

pw2015_t2_relationship$pastrelationship3_relationship_new <- sub('_NA*','', pw2015_t2_relationship$pastrelationship3_relationship_new)
pw2015_t2_relationship$pastrelationship3_relationship_new <- sub('NA_*','', pw2015_t2_relationship$pastrelationship3_relationship_new)
pw2015_t2_relationship$pastrelationship3_relationship_new[pw2015_t2_relationship$pastrelationship3_relationship_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship3_relationship_new)) %>% separate(pastrelationship3_relationship_new, c("pastrelationship3_relationship_new","pastrelationship8_relationship_new"), sep = "_")

pw2015_t2_relationship$pastrelationship3_duration_new <- sub('_NA*','', pw2015_t2_relationship$pastrelationship3_duration_new)
pw2015_t2_relationship$pastrelationship3_duration_new <- sub('NA_*','', pw2015_t2_relationship$pastrelationship3_duration_new)
pw2015_t2_relationship$pastrelationship3_duration_new[pw2015_t2_relationship$pastrelationship3_duration_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship3_duration_new)) %>% separate(pastrelationship3_duration_new, c("pastrelationship3_duration_new","pastrelationship8_duration_new"), sep = "_")
```

Past relationship 4
```{r}
pw2015_t2_relationship$pastrelationship4_kind <- sub('_NA*','', pw2015_t2_relationship$pastrelationship4_kind)
pw2015_t2_relationship$pastrelationship4_kind <- sub('NA_*','', pw2015_t2_relationship$pastrelationship4_kind)
pw2015_t2_relationship$pastrelationship4_kind[pw2015_t2_relationship$pastrelationship4_kind==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship4_kind)) %>% separate(pastrelationship4_kind, c("pastrelationship4_kind","pastrelationship9_kind"), sep = "_")

pw2015_t2_relationship$pastrelationship4_start <- sub('_NA*','', pw2015_t2_relationship$pastrelationship4_start)
pw2015_t2_relationship$pastrelationship4_start <- sub('NA_*','', pw2015_t2_relationship$pastrelationship4_start)
pw2015_t2_relationship$pastrelationship4_start[pw2015_t2_relationship$pastrelationship4_start==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship4_start)) %>% separate(pastrelationship4_start, c("pastrelationship4_start","pastrelationship9_start"), sep = "_")

pw2015_t2_relationship$pastrelationship4_end <- sub('_NA*','', pw2015_t2_relationship$pastrelationship4_end)
pw2015_t2_relationship$pastrelationship4_end <- sub('NA_*','', pw2015_t2_relationship$pastrelationship4_end)
pw2015_t2_relationship$pastrelationship4_end[pw2015_t2_relationship$pastrelationship4_end==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship4_end)) %>% separate(pastrelationship4_end, c("pastrelationship4_end","pastrelationship9_end"), sep = "_")

pw2015_t2_relationship$pastrelationship4_name <- sub('_NA*','', pw2015_t2_relationship$pastrelationship4_name)
pw2015_t2_relationship$pastrelationship4_name <- sub('NA_*','', pw2015_t2_relationship$pastrelationship4_name)
pw2015_t2_relationship$pastrelationship4_name[pw2015_t2_relationship$pastrelationship4_name==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship4_name)) %>% separate(pastrelationship4_name, c("pastrelationship4_name","pastrelationship9_name"), sep = "_")

pw2015_t2_relationship$pastrelationship4_name_new <- sub('_NA*','', pw2015_t2_relationship$pastrelationship4_name_new)
pw2015_t2_relationship$pastrelationship4_name_new <- sub('NA_*','', pw2015_t2_relationship$pastrelationship4_name_new)
pw2015_t2_relationship$pastrelationship4_name_new[pw2015_t2_relationship$pastrelationship4_name_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship4_name_new)) %>% separate(pastrelationship4_name_new, c("pastrelationship4_name_new","pastrelationship9_name_new"), sep = "_")

pw2015_t2_relationship$pastrelationship4_age_new <- sub('_NA*','', pw2015_t2_relationship$pastrelationship4_age_new)
pw2015_t2_relationship$pastrelationship4_age_new <- sub('NA_*','', pw2015_t2_relationship$pastrelationship4_age_new)
pw2015_t2_relationship$pastrelationship4_age_new[pw2015_t2_relationship$pastrelationship4_age_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship4_age_new)) %>% separate(pastrelationship4_age_new, c("pastrelationship4_age_new","pastrelationship9_age_new"), sep = "_")

pw2015_t2_relationship$pastrelationship4_height_new <- sub('_NA*','', pw2015_t2_relationship$pastrelationship4_height_new)
pw2015_t2_relationship$pastrelationship4_height_new <- sub('NA_*','', pw2015_t2_relationship$pastrelationship4_height_new)
pw2015_t2_relationship$pastrelationship4_height_new[pw2015_t2_relationship$pastrelationship4_height_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship4_height_new)) %>% separate(pastrelationship4_height_new, c("pastrelationship4_height_new","pastrelationship9_height_new"), sep = "_")

pw2015_t2_relationship$pastrelationship4_relationship_new <- sub('_NA*','', pw2015_t2_relationship$pastrelationship4_relationship_new)
pw2015_t2_relationship$pastrelationship4_relationship_new <- sub('NA_*','', pw2015_t2_relationship$pastrelationship4_relationship_new)
pw2015_t2_relationship$pastrelationship4_relationship_new[pw2015_t2_relationship$pastrelationship4_relationship_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship4_relationship_new)) %>% separate(pastrelationship4_relationship_new, c("pastrelationship4_relationship_new","pastrelationship9_relationship_new"), sep = "_")

pw2015_t2_relationship$pastrelationship4_duration_new <- sub('_NA*','', pw2015_t2_relationship$pastrelationship4_duration_new)
pw2015_t2_relationship$pastrelationship4_duration_new <- sub('NA_*','', pw2015_t2_relationship$pastrelationship4_duration_new)
pw2015_t2_relationship$pastrelationship4_duration_new[pw2015_t2_relationship$pastrelationship4_duration_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship4_duration_new)) %>% separate(pastrelationship4_duration_new, c("pastrelationship4_duration_new","pastrelationship9_duration_new"), sep = "_")
```

Past relationship 5
```{r}
pw2015_t2_relationship$pastrelationship5_kind <- sub('_NA*','', pw2015_t2_relationship$pastrelationship5_kind)
pw2015_t2_relationship$pastrelationship5_kind <- sub('NA_*','', pw2015_t2_relationship$pastrelationship5_kind)
pw2015_t2_relationship$pastrelationship5_kind[pw2015_t2_relationship$pastrelationship5_kind==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship5_kind)) %>% separate(pastrelationship5_kind, c("pastrelationship5_kind","pastrelationship10_kind"), sep = "_")

pw2015_t2_relationship$pastrelationship5_start <- sub('_NA*','', pw2015_t2_relationship$pastrelationship5_start)
pw2015_t2_relationship$pastrelationship5_start <- sub('NA_*','', pw2015_t2_relationship$pastrelationship5_start)
pw2015_t2_relationship$pastrelationship5_start[pw2015_t2_relationship$pastrelationship5_start==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship5_start)) %>% separate(pastrelationship5_start, c("pastrelationship5_start","pastrelationship10_start"), sep = "_")

pw2015_t2_relationship$pastrelationship5_end <- sub('_NA*','', pw2015_t2_relationship$pastrelationship5_end)
pw2015_t2_relationship$pastrelationship5_end <- sub('NA_*','', pw2015_t2_relationship$pastrelationship5_end)
pw2015_t2_relationship$pastrelationship5_end[pw2015_t2_relationship$pastrelationship5_end==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship5_end)) %>% separate(pastrelationship5_end, c("pastrelationship5_end","pastrelationship10_end"), sep = "_")

pw2015_t2_relationship$pastrelationship5_name <- sub('_NA*','', pw2015_t2_relationship$pastrelationship5_name)
pw2015_t2_relationship$pastrelationship5_name <- sub('NA_*','', pw2015_t2_relationship$pastrelationship5_name)
pw2015_t2_relationship$pastrelationship5_name[pw2015_t2_relationship$pastrelationship5_name==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship5_name)) %>% separate(pastrelationship5_name, c("pastrelationship5_name","pastrelationship10_name"), sep = "_")

pw2015_t2_relationship$pastrelationship5_name_new <- sub('_NA*','', pw2015_t2_relationship$pastrelationship5_name_new)
pw2015_t2_relationship$pastrelationship5_name_new <- sub('NA_*','', pw2015_t2_relationship$pastrelationship5_name_new)
pw2015_t2_relationship$pastrelationship5_name_new[pw2015_t2_relationship$pastrelationship5_name_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship5_name_new)) %>% separate(pastrelationship5_name_new, c("pastrelationship5_name_new","pastrelationship10_name_new"), sep = "_")

pw2015_t2_relationship$pastrelationship5_age_new <- sub('_NA*','', pw2015_t2_relationship$pastrelationship5_age_new)
pw2015_t2_relationship$pastrelationship5_age_new <- sub('NA_*','', pw2015_t2_relationship$pastrelationship5_age_new)
pw2015_t2_relationship$pastrelationship5_age_new[pw2015_t2_relationship$pastrelationship5_age_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship5_age_new)) %>% separate(pastrelationship5_age_new, c("pastrelationship5_age_new","pastrelationship10_age_new"), sep = "_")

pw2015_t2_relationship$pastrelationship5_height_new <- sub('_NA*','', pw2015_t2_relationship$pastrelationship5_height_new)
pw2015_t2_relationship$pastrelationship5_height_new <- sub('NA_*','', pw2015_t2_relationship$pastrelationship5_height_new)
pw2015_t2_relationship$pastrelationship5_height_new[pw2015_t2_relationship$pastrelationship5_height_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship5_height_new)) %>% separate(pastrelationship5_height_new, c("pastrelationship5_height_new","pastrelationship10_height_new"), sep = "_")

pw2015_t2_relationship$pastrelationship5_relationship_new <- sub('_NA*','', pw2015_t2_relationship$pastrelationship5_relationship_new)
pw2015_t2_relationship$pastrelationship5_relationship_new <- sub('NA_*','', pw2015_t2_relationship$pastrelationship5_relationship_new)
pw2015_t2_relationship$pastrelationship5_relationship_new[pw2015_t2_relationship$pastrelationship5_relationship_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship5_relationship_new)) %>% separate(pastrelationship5_relationship_new, c("pastrelationship5_relationship_new","pastrelationship10_relationship_new"), sep = "_")

pw2015_t2_relationship$pastrelationship5_duration_new <- sub('_NA*','', pw2015_t2_relationship$pastrelationship5_duration_new)
pw2015_t2_relationship$pastrelationship5_duration_new <- sub('NA_*','', pw2015_t2_relationship$pastrelationship5_duration_new)
pw2015_t2_relationship$pastrelationship5_duration_new[pw2015_t2_relationship$pastrelationship5_duration_new==""]  <- NA 
pw2015_t2_relationship <- pw2015_t2_relationship %>% group_by(grepl('_', pastrelationship5_duration_new)) %>% separate(pastrelationship5_duration_new, c("pastrelationship5_duration_new","pastrelationship10_duration_new"), sep = "_")
```

#Add created_singles_t2 variable for right censored data
 
```{r}
 pw2015_t2_relationship$created.m = as.Date(pw2015_t2_relationship$created.m, origin = "1970-01-01")
pw2015_t2_relationship$created.f = as.Date(pw2015_t2_relationship$created.f, origin = "1970-01-01")

pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  created_singles_t2 = if_else(is.na(created.f), created.m, created.f))
```

Bring data with new variables in proper order
```{r}
#Delete grouping-variable "grepl"
pw2015_t2_relationship <- pw2015_t2_relationship %>% select( -starts_with("grepl"))

pw2015_t2_relationship <- pw2015_t2_relationship %>% select(-relationship_m, -relationship_f, -pastrelationship1_m,
                          -pastrelationship1_f, -pastrelationship2_m, -pastrelationship2_f, -pastrelationship3_m,
                          -pastrelationship3_f, -pastrelationship4_m, -pastrelationship4_f, -pastrelationship5_m,
                          -pastrelationship5_f, -pastrelationship6_m, -pastrelationship6_f,
                          -pastrelationshipmore_m, -pastrelationshipmore_f)

pw2015_t2_relationship <- pw2015_t2_relationship %>% select(session, short_session, t2_relationship, 
                             created_singles_t2,  t2_pastrelationship, 
                             starts_with("relationship"), starts_with("pastrelationship"))


```



```{r}
pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
    relationship1.name = relationship1_name, relationship1.name_new = relationship1_name_new, 
    relationship1.age_new = relationship1_age_new, relationship1.height_new = relationship1_height_new,
    relationship1.relationship_new = relationship1_relationship_new, relationship1.duration_new = 
    relationship1_duration_new, relationship1.kind = relationship1_kind, relationship1.start = relationship1_start, 
    relationship2.name = relationship2_name, relationship2.name_new = relationship2_name_new, 
    relationship2.age_new = relationship2_age_new, relationship2.height_new = relationship2_height_new, 
    relationship2.relationship_new = relationship2_relationship_new, relationship2.duration_new = 
    relationship2_duration_new, relationship2.kind = relationship2_kind, relationship2.start = relationship2_start, 
    relationship3.name = relationship3_name, relationship3.name_new = relationship3_name_new, 
    relationship3.age_new = relationship3_age_new, relationship3.height_new = relationship3_height_new,
    relationship3.relationship_new = relationship3_relationship_new, relationship3.duration_new = 
    relationship3_duration_new, relationship3.kind = relationship3_kind, relationship3.start = relationship3_start, 
    relationship4.name = relationship4_name, relationship4.name_new = relationship4_name_new, 
    relationship4.age_new = relationship4_age_new, relationship4.height_new = relationship4_height_new, 
    relationship4.relationship_new = relationship4_relationship_new, relationship4.duration_new = 
    relationship4_duration_new, relationship4.kind = relationship4_kind, relationship4.start = relationship4_start,            relationship5.name = relationship5_name, relationship5.name_new = relationship5_name_new, 
    relationship5.age_new = relationship5_age_new, relationship5.height_new = relationship5_height_new,
    relationship5.relationship_new = relationship5_relationship_new, relationship5.duration_new = 
    relationship5_duration_new, relationship5.kind = relationship5_kind, relationship5.start = relationship5_start, 
    relationship6.name = relationship6_name, relationship6.name_new = relationship6_name_new, 
    relationship6.age_new = relationship6_age_new, relationship6.height_new = relationship6_height_new, 
    relationship6.relationship_new = relationship6_relationship_new, relationship6.duration_new = 
    relationship6_duration_new, relationship6.kind = relationship6_kind, relationship6.start = relationship6_start, 
    pastrelationship1.name = pastrelationship1_name, pastrelationship1.name_new = pastrelationship1_name_new,  
    pastrelationship1.age_new = pastrelationship1_age_new, pastrelationship1.height_new = 
    pastrelationship1_height_new, pastrelationship1.relationship_new = pastrelationship1_relationship_new, 
    pastrelationship1.duration_new = pastrelationship1_duration_new, pastrelationship1.kind = 
    pastrelationship1_kind, pastrelationship1.start = pastrelationship1_start, pastrelationship1.end = 
    pastrelationship1_end,
    pastrelationship2.name = pastrelationship2_name, pastrelationship2.name_new = pastrelationship2_name_new,  
    pastrelationship2.age_new = pastrelationship2_age_new, pastrelationship2.height_new = 
    pastrelationship2_height_new, pastrelationship2.relationship_new = pastrelationship2_relationship_new, 
    pastrelationship2.duration_new = pastrelationship2_duration_new, pastrelationship2.kind = 
    pastrelationship2_kind, pastrelationship2.start = pastrelationship2_start, pastrelationship2.end = 
    pastrelationship2_end,
    pastrelationship3.name = pastrelationship3_name, pastrelationship3.name_new = pastrelationship3_name_new,  
    pastrelationship3.age_new = pastrelationship3_age_new, pastrelationship3.height_new = 
    pastrelationship3_height_new, pastrelationship3.relationship_new = pastrelationship3_relationship_new, 
    pastrelationship3.duration_new = pastrelationship3_duration_new, pastrelationship3.kind = 
    pastrelationship3_kind, pastrelationship3.start = pastrelationship3_start, pastrelationship3.end = 
    pastrelationship3_end,
    pastrelationship4.name = pastrelationship4_name, pastrelationship4.name_new = pastrelationship4_name_new,  
    pastrelationship4.age_new = pastrelationship4_age_new, pastrelationship4.height_new = 
    pastrelationship4_height_new, pastrelationship4.relationship_new = pastrelationship4_relationship_new, 
    pastrelationship4.duration_new = pastrelationship4_duration_new, pastrelationship4.kind = 
    pastrelationship4_kind, pastrelationship4.start = pastrelationship4_start, pastrelationship4.end = 
    pastrelationship4_end,
    pastrelationship5.name = pastrelationship5_name, pastrelationship5.name_new = pastrelationship5_name_new,  
    pastrelationship5.age_new = pastrelationship5_age_new, pastrelationship5.height_new = 
    pastrelationship5_height_new, pastrelationship5.relationship_new = pastrelationship5_relationship_new, 
    pastrelationship5.duration_new = pastrelationship5_duration_new, pastrelationship5.kind = 
    pastrelationship5_kind, pastrelationship5.start = pastrelationship5_start, pastrelationship5.end = 
    pastrelationship5_end,
    pastrelationship6.name = pastrelationship6_name, pastrelationship6.name_new = pastrelationship6_name_new,  
    pastrelationship6.age_new = pastrelationship6_age_new, pastrelationship6.height_new = 
    pastrelationship6_height_new, pastrelationship6.relationship_new = pastrelationship6_relationship_new, 
    pastrelationship6.duration_new = pastrelationship6_duration_new, pastrelationship6.kind = 
    pastrelationship6_kind, pastrelationship6.start = pastrelationship6_start, pastrelationship6.end = 
    pastrelationship6_end,
    pastrelationship7.name = pastrelationship7_name, pastrelationship7.name_new = pastrelationship7_name_new,  
    pastrelationship7.age_new = pastrelationship7_age_new, pastrelationship7.height_new = 
    pastrelationship7_height_new, pastrelationship7.relationship_new = pastrelationship7_relationship_new, 
    pastrelationship7.duration_new = pastrelationship7_duration_new, pastrelationship7.kind = 
    pastrelationship7_kind, pastrelationship7.start = pastrelationship7_start, pastrelationship7.end = 
    pastrelationship7_end,
    pastrelationship8.name = pastrelationship8_name, pastrelationship8.name_new = pastrelationship8_name_new,  
    pastrelationship8.age_new = pastrelationship8_age_new, pastrelationship8.height_new = 
    pastrelationship8_height_new, pastrelationship8.relationship_new = pastrelationship8_relationship_new, 
    pastrelationship8.duration_new = pastrelationship8_duration_new, pastrelationship8.kind = 
    pastrelationship8_kind, pastrelationship8.start = pastrelationship8_start, pastrelationship8.end = 
    pastrelationship8_end,
    pastrelationship9.name = pastrelationship9_name, pastrelationship9.name_new = pastrelationship9_name_new,  
    pastrelationship9.age_new = pastrelationship9_age_new, pastrelationship9.height_new = 
    pastrelationship9_height_new, pastrelationship9.relationship_new = pastrelationship9_relationship_new, 
    pastrelationship9.duration_new = pastrelationship9_duration_new, pastrelationship9.kind = 
    pastrelationship9_kind, pastrelationship9.start = pastrelationship9_start, pastrelationship9.end = 
    pastrelationship9_end,
    pastrelationship10.name = pastrelationship10_name, pastrelationship10.name_new = pastrelationship10_name_new,  
    pastrelationship10.age_new = pastrelationship10_age_new, pastrelationship10.height_new = 
    pastrelationship10_height_new, pastrelationship10.relationship_new = pastrelationship10_relationship_new, 
    pastrelationship10.duration_new = pastrelationship10_duration_new, pastrelationship10.kind = 
    pastrelationship10_kind, pastrelationship10.start = pastrelationship10_start, pastrelationship10.end = 
    pastrelationship10_end)

pw2015_t2_relationship <- pw2015_t2_relationship %>% select( -starts_with("relationship1_"),  
                        -starts_with("relationship2_"),  -starts_with("relationship3_"),  -starts_with("relationship4_"),  
                        -starts_with("relationship5_"),  -starts_with("relationship6_"),-starts_with("pastrelationship1_"),
                        -starts_with("pastrelationship2_"), -starts_with("pastrelationship3_"), 
                        -starts_with("pastrelationship4_"), -starts_with("pastrelationship5_"), 
                        -starts_with("pastrelationship6_"), -starts_with("pastrelationship7_"),
                        -starts_with("pastrelationship8_"), -starts_with("pastrelationship9_"),
                        -starts_with("pastrelationship10_"))
                                                                  
pw2015_t2_relationship <- pw2015_t2_relationship %>% mutate(
  "relationship1.end" = NA, "relationship2.end" = NA, "relationship3.end" = NA,
  "relationship4.end" = NA, "relationship5.end" = NA, "relationship6.end" = NA
)


pw2015_t2_relationship <- as.data.frame(pw2015_t2_relationship)
```


Reshape data : Variables must be listed in alphabetical order
```{r}
pw2015_t2_relationship <- reshape(pw2015_t2_relationship, direction='long', 
        varying=c('relationship1.age_new', 'relationship1.duration_new', 'relationship1.end',
                  'relationship1.height_new', 'relationship1.kind', 'relationship1.name',
                  'relationship1.name_new', 'relationship1.relationship_new', 'relationship1.start',
                  
                  'relationship2.age_new', 'relationship2.duration_new', 'relationship2.end',
                  'relationship2.height_new', 'relationship2.kind', 'relationship2.name',
                  'relationship2.name_new', 'relationship2.relationship_new', 'relationship2.start',
                  
                  'relationship3.age_new', 'relationship3.duration_new', 'relationship3.end',
                  'relationship3.height_new', 'relationship3.kind', 'relationship3.name',
                  'relationship3.name_new', 'relationship3.relationship_new', 'relationship3.start',
                 
                  'relationship4.age_new', 'relationship4.duration_new', 'relationship4.end',
                  'relationship4.height_new', 'relationship4.kind', 'relationship4.name',
                  'relationship4.name_new', 'relationship4.relationship_new', 'relationship4.start',
                  
                  'relationship5.age_new', 'relationship5.duration_new', 'relationship5.end',
                  'relationship5.height_new', 'relationship5.kind', 'relationship5.name',
                  'relationship5.name_new', 'relationship5.relationship_new', 'relationship5.start',
                  
                  'relationship6.age_new', 'relationship6.duration_new', 'relationship6.end',
                  'relationship6.height_new', 'relationship6.kind', 'relationship6.name',
                  'relationship6.name_new', 'relationship6.relationship_new', 'relationship6.start',
                  
                  'pastrelationship1.age_new', 'pastrelationship1.duration_new', 'pastrelationship1.end',
                  'pastrelationship1.height_new', 'pastrelationship1.kind', 'pastrelationship1.name',
                  'pastrelationship1.name_new', 'pastrelationship1.relationship_new', 'pastrelationship1.start',
                  
                  'pastrelationship2.age_new', 'pastrelationship2.duration_new', 'pastrelationship2.end',
                  'pastrelationship2.height_new', 'pastrelationship2.kind', 'pastrelationship2.name',
                  'pastrelationship2.name_new', 'pastrelationship2.relationship_new', 'pastrelationship2.start',
                  
                  'pastrelationship3.age_new', 'pastrelationship3.duration_new', 'pastrelationship3.end',
                  'pastrelationship3.height_new', 'pastrelationship3.kind', 'pastrelationship3.name',
                  'pastrelationship3.name_new', 'pastrelationship3.relationship_new', 'pastrelationship3.start',
                 
                  'pastrelationship4.age_new', 'pastrelationship4.duration_new', 'pastrelationship4.end',
                  'pastrelationship4.height_new', 'pastrelationship4.kind', 'pastrelationship4.name',
                  'pastrelationship4.name_new', 'pastrelationship4.relationship_new', 'pastrelationship4.start',
                 
                  'pastrelationship5.age_new', 'pastrelationship5.duration_new', 'pastrelationship5.end',
                  'pastrelationship5.height_new', 'pastrelationship5.kind', 'pastrelationship5.name',
                  'pastrelationship5.name_new', 'pastrelationship5.relationship_new', 'pastrelationship5.start',
                 
                  'pastrelationship6.age_new', 'pastrelationship6.duration_new', 'pastrelationship6.end',
                  'pastrelationship6.height_new', 'pastrelationship6.kind', 'pastrelationship6.name',
                  'pastrelationship6.name_new', 'pastrelationship6.relationship_new', 'pastrelationship6.start',
                 
                  'pastrelationship7.age_new', 'pastrelationship7.duration_new', 'pastrelationship7.end',
                  'pastrelationship7.height_new', 'pastrelationship7.kind', 'pastrelationship7.name',
                  'pastrelationship7.name_new', 'pastrelationship7.relationship_new', 'pastrelationship7.start',
                  
                  'pastrelationship8.age_new', 'pastrelationship8.duration_new', 'pastrelationship8.end',
                  'pastrelationship8.height_new', 'pastrelationship8.kind', 'pastrelationship8.name',
                  'pastrelationship8.name_new', 'pastrelationship8.relationship_new', 'pastrelationship8.start',
                  
                  'pastrelationship9.age_new', 'pastrelationship9.duration_new', 'pastrelationship9.end',
                  'pastrelationship9.height_new', 'pastrelationship9.kind', 'pastrelationship9.name',
                  'pastrelationship9.name_new', 'pastrelationship9.relationship_new', 'pastrelationship9.start',
                  
                  'pastrelationship10.age_new', 'pastrelationship10.duration_new', 'pastrelationship10.end',
                  'pastrelationship10.height_new', 'pastrelationship10.kind', 'pastrelationship10.name',
                  'pastrelationship10.name_new', 'pastrelationship10.relationship_new', 'pastrelationship10.start'),
             timevar='var',
        times=c('relationship1', 'relationship2', 'relationship3', 'relationship4', 'relationship5', 'relationship6',
                'pastrelationship1', 'pastrelationship2', 'pastrelationship3', 'pastrelationship4', 'pastrelationship5',
                'pastrelationship6', 'pastrelationship7', 'pastrelationship8', 'pastrelationship9', 'pastrelationship10'),
        v.names=c('age_new', 'duration_new', 'end', 'height_new', 'kind', 'name', 'name_new', 'relationship_new', 'start'),
        idvar='short_session')
```


Change format: 1 relationship per row, or if single whole time one row with only NAs
```{r}
#Dataframe with only participants who had at least one relationship between t1 and t2
pw2015_t2_relationship_1 <- pw2015_t2_relationship %>%  filter(t2_relationship != 0 | t2_pastrelationship != 0) %>%
            filter(!is.na(age_new)| !is.na(duration_new) | !is.na(end) | !is.na(height_new) | !is.na(kind) | 
            !is.na(name) | !is.na(name_new) | !is.na(relationship_new) | !is.na(start)) 

#Dataframe with only participants who had NONE relationship between t1 and t2
pw2015_t2_relationship_2 <- pw2015_t2_relationship %>% 
              filter(t2_relationship == 0 & t2_pastrelationship == 0) %>% filter(var == "relationship1") %>%
  mutate(var = "Single")


pw2015_t2_relationship <- bind_rows(pw2015_t2_relationship_1, pw2015_t2_relationship_2)  

pw2015_t2_relationship <-  pw2015_t2_relationship %>% 
  group_by(session) %>% # we want to look at relationships specifically for each person
  arrange(session, start) %>% #these relationships should be arranged starting with the earliest
                                   # starting data
  mutate(relationship_number = n(), # counting total number of relationships
         relationship_order = row_number()) %>% # determining relationship order
  ungroup()
```

Add variable lasting: so dataframes of t2 and t3 can be joined
```{r}
pw2015_t2_relationship <- pw2015_t2_relationship %>%
  mutate(lasting = var)
    
pw2015_t2_relationship$lasting[pw2015_t2_relationship$lasting == "relationship1"] <- 1
pw2015_t2_relationship$lasting[pw2015_t2_relationship$lasting == "relationship2"] <- 1
pw2015_t2_relationship$lasting[pw2015_t2_relationship$lasting == "relationship3"] <- 1
pw2015_t2_relationship$lasting[pw2015_t2_relationship$lasting == "relationship4"] <- 1
pw2015_t2_relationship$lasting[pw2015_t2_relationship$lasting == "relationship5"] <- 1
pw2015_t2_relationship$lasting[pw2015_t2_relationship$lasting == "relationship6"] <- 1
 
pw2015_t2_relationship$lasting[pw2015_t2_relationship$lasting == "pastrelationship1"] <- 2
pw2015_t2_relationship$lasting[pw2015_t2_relationship$lasting == "pastrelationship2"] <- 2
pw2015_t2_relationship$lasting[pw2015_t2_relationship$lasting == "pastrelationship3"] <- 2
pw2015_t2_relationship$lasting[pw2015_t2_relationship$lasting == "pastrelationship4"] <- 2
pw2015_t2_relationship$lasting[pw2015_t2_relationship$lasting == "pastrelationship5"] <- 2
pw2015_t2_relationship$lasting[pw2015_t2_relationship$lasting == "pastrelationship6"] <- 2
pw2015_t2_relationship$lasting[pw2015_t2_relationship$lasting == "pastrelationship7"] <- 2
pw2015_t2_relationship$lasting[pw2015_t2_relationship$lasting == "pastrelationship8"] <- 2
pw2015_t2_relationship$lasting[pw2015_t2_relationship$lasting == "pastrelationship9"] <- 2
pw2015_t2_relationship$lasting[pw2015_t2_relationship$lasting == "pastrelationship10"] <- 2

pw2015_t2_relationship$lasting[pw2015_t2_relationship$lasting == "Single"] <- 0

pw2015_t2_relationship$lasting <- as.integer(pw2015_t2_relationship$lasting)

```

Add sex variable for partners
```{r}
pw2015_t2_relationship <- pw2015_t2_relationship %>%
  mutate(sex_partner = var)

# 1 = female, 2 = male

#based on relationship variables from t2_male & t2_female partner sheets:

pw2015_t2_relationship$sex_partner[pw2015_t2_relationship$sex_partner == "relationship1"] <- 2
pw2015_t2_relationship$sex_partner[pw2015_t2_relationship$sex_partner == "relationship2"] <- 2
pw2015_t2_relationship$sex_partner[pw2015_t2_relationship$sex_partner == "relationship3"] <- 2
pw2015_t2_relationship$sex_partner[pw2015_t2_relationship$sex_partner == "relationship4"] <- 1
pw2015_t2_relationship$sex_partner[pw2015_t2_relationship$sex_partner == "relationship5"] <- 1
pw2015_t2_relationship$sex_partner[pw2015_t2_relationship$sex_partner == "relationship6"] <- 1
 
pw2015_t2_relationship$sex_partner[pw2015_t2_relationship$sex_partner == "pastrelationship1"] <- 2
pw2015_t2_relationship$sex_partner[pw2015_t2_relationship$sex_partner == "pastrelationship2"] <- 2
pw2015_t2_relationship$sex_partner[pw2015_t2_relationship$sex_partner == "pastrelationship3"] <- 2
pw2015_t2_relationship$sex_partner[pw2015_t2_relationship$sex_partner == "pastrelationship4"] <- 2
pw2015_t2_relationship$sex_partner[pw2015_t2_relationship$sex_partner == "pastrelationship5"] <- 2
pw2015_t2_relationship$sex_partner[pw2015_t2_relationship$sex_partner == "pastrelationship6"] <- 1
pw2015_t2_relationship$sex_partner[pw2015_t2_relationship$sex_partner == "pastrelationship7"] <- 1
pw2015_t2_relationship$sex_partner[pw2015_t2_relationship$sex_partner == "pastrelationship8"] <- 1
pw2015_t2_relationship$sex_partner[pw2015_t2_relationship$sex_partner == "pastrelationship9"] <- 1
pw2015_t2_relationship$sex_partner[pw2015_t2_relationship$sex_partner == "pastrelationship10"] <- 1

pw2015_t2_relationship$sex_partner <- as.integer(pw2015_t2_relationship$sex_partner)
```

# Liars in t2
```{r}
# Get starting date of t1
starting_date = pw2015_demographie %>% select(session, created_t1 = created)


#t2
non_singles_t2 = left_join(pw2015_t2_relationship, starting_date, by = "session") %>%
  mutate(check_liars = ifelse(as.Date(created_t1) < as.Date(start), 0, 1)) %>%
  filter(check_liars == 1)

non_singles_t2_session = non_singles_t2 %>% filter(!duplicated(session))


count(non_singles_t2_session)
table(non_singles_t2_session$kind)  

non_singles_vector_t2 = as.data.frame(non_singles_t2_session$session)

# t3
non_singles_t3 = left_join(pw2015_t3_partners, starting_date, by = "session") %>%
  mutate(check_liars = ifelse(as.Date(created_t1) < as.Date(F_t3_start), 0, 1)) %>%
  filter(check_liars == 1)

non_singles_vector_t3 = as.data.frame(non_singles_t3$session)

count(non_singles_vector_t3)

non_singles_vector = bind_rows(non_singles_vector_t2, non_singles_vector_t3) %>%
  mutate(session = ifelse(is.na(`non_singles_t2_session$session`),
                          `non_singles_t3$session`,
                          `non_singles_t2_session$session`))

non_singles_vector_unique = non_singles_vector %>% filter(!duplicated(session))
non_singles_vector_unique = non_singles_vector_unique$`non_singles_t2_session$session`
length(non_singles_vector_unique)
```

# Data t3

```{r}
# singles create t3
pw2015_t3_singles$created_singles_t3 = as.Date(pw2015_t3_singles$created)
pw2015_t3_partners$F_t3_start = as.Date(pw2015_t3_partners$F_t3_start)
# first we have to combine singles and partnered reports
pw2015_t3_partners = bind_rows(pw2015_t3_partners, pw2015_t3_singles)
# Relationships for t3 are already stored in the right format. We just have to count total relationship for every person and order and count relationships.
pw2015_t3_partners = pw2015_t3_partners %>% 
  group_by(session) %>% # we want to look at relationships specifically for each person
  arrange(session, F_t3_start) %>% #these relationships should be arranged starting with the earliest
                                   # starting data
  mutate(relationship_number = n(), # counting total number of relationships
         relationship_order = row_number(), # determining relationship order
         kind = ifelse(!is.na(F_t3_kind), F_t3_kind, F_t3_past_kind)) %>%
  mutate(relationship_number = ifelse(is.na(F_t3_start), 0, relationship_number),
         relationship_order = ifelse(is.na(F_t3_start), 0, relationship_order),
         F_t3_lasting = ifelse(is.na(F_t3_start), 0, F_t3_lasting))

pw2015_t3_partners = pw2015_t3_partners %>%
  mutate(F_t3_start = if_else(is.na(F_t3_start), created_singles_t3, F_t3_start)) %>%
  filter(!is.na(F_t3_start)) # exclude weird cases - check this if you have time
```

# Relationships at t2 and t3
```{r}
# Remove relationships from t2 that have information in t3 before merging based on names
# Form Variable including name for t3
pw2015_t3_partners = pw2015_t3_partners %>%
  mutate(F_t3_description = F_t3_name) %>%
# Name:
  separate(F_t3_description,
           into = c("F_t3_name"),
           sep = " ",
           remove = FALSE,
           extra = "drop") %>%
  mutate(F_t3_name = as.character(F_t3_name),
         F_t3_name_new = as.character(F_t3_name_new),
         F_t3_name = ifelse(F_t3_name == "Kommt", F_t3_name_new, F_t3_name)) %>%
  mutate(F_t3_sex_old = ifelse(F_t3_description %contains% "<U+2642>", 2,
                               ifelse(F_t3_description %contains% "<U+2640>", 1, NA)),
         F_t3_sex = ifelse(is.na(F_t3_sex_old), F_t3_sex_new, F_t3_sex_old))

# Remove double singles
double_singles = pw2015_t2_relationship %>%
  filter(lasting == 0) %>%
  filter(session %in% pw2015_t3_partners$session)

pw2015_t2_relationship = pw2015_t2_relationship %>%
  filter(!(session %in% double_singles$session))

# Remove duble relationships
# Form matching variable for t2 and t3
double_relationships_t3 = pw2015_t3_partners %>%
  filter(!is.na(F_t3_name)) %>%
  mutate(session_partnername = paste0(session, "_", F_t3_name))

double_relationships_t2 = pw2015_t2_relationship %>%
  mutate(name = ifelse(name == "Kommt nicht in dieser Liste vor", name_new, name)) %>%
  filter(!is.na(name), lasting == 1) %>%
  mutate(session_partnername = paste0(session, "_", name)) %>%
  filter(session_partnername %in% double_relationships_t3$session_partnername)


pw2015_t2_relationship = pw2015_t2_relationship %>%
  filter(!(short_session == "hLWe0YPc9g" & is.na(name))) %>%
  filter(!(short_session == "ocQ1e9cRL4" & is.na(name))) %>%
  mutate(name = ifelse(name == "Kommt nicht in dieser Liste vor", name_new, name),
         name = ifelse(is.na(name), name_new, name),
         session_partnername = ifelse(lasting == 1, paste0(session, "_", name), NA)) %>%
  filter(!(session_partnername %in% double_relationships_t2$session_partnername))

############################ Maybe rename data -> t2_test / t3_test
t2_test <- pw2015_t2_relationship %>% ungroup() %>% select(session, short_session, start, end, created_singles_t2,
                                                           kind, lasting, name, sex_partner)

t2_test$start = as.Date(t2_test$start)

pw2015_t3_partners = pw2015_t3_partners %>%
  mutate(kind = ifelse(is.na(F_t3_kind), F_t3_past_kind, F_t3_kind))

t3_test <- pw2015_t3_partners %>% ungroup() %>% select(session, short_session, created, F_t3_start,
                                                       F_t3_sex,F_t3_end, kind, F_t3_lasting,
                                                       F_t3_name)

t3_test <- t3_test %>% rename(
                      start = F_t3_start, 
                      end = F_t3_end,
                      lasting = F_t3_lasting,
                      name = F_t3_name,
                      sex_partner = F_t3_sex, 
                      t3_created = created) %>%
  mutate(kind = as.character(kind))
                      

t2_t3_relationship <- bind_rows(t2_test, t3_test)

# Remove weird cases before analysis
t2_t3_relationship = t2_t3_relationship %>% filter(!(lasting != 0 & is.na(kind)))

t2_t3_relationship_exclusive = t2_t3_relationship %>% 
  filter(kind == 1 | is.na(kind)) %>%
  group_by(session) %>% # we want to look at relationships specifically for each person
  arrange(session, start) %>% #these relationships should be arranged starting with the earliest
                                   # starting data
  mutate(relationship_number = n(), # counting total number of relationships
         relationship_order = row_number(),
         # lasting = ifelse(is.na(start), 0, lasting), done earliere (hopefully)
         relationship_order = ifelse(is.na(start), 0, relationship_order),
         relationship_number = ifelse(is.na(start), 0, relationship_number)) 


t2_t3_relationship = t2_t3_relationship %>% 
  group_by(session) %>% # we want to look at relationships specifically for each person
  arrange(session, start) %>% #these relationships should be arranged starting with the earliest
                                   # starting data
  mutate(relationship_number = n(), # counting total number of relationships
         relationship_order = row_number(),
         # lasting = ifelse(is.na(start), 0, lasting), done earliere (hopefully)
         relationship_order = ifelse(is.na(start), 0, relationship_order),
         relationship_number = ifelse(is.na(start), 0, relationship_number)) 

```

# Merge Data

```{r}
#create selected data sheets
#pw2015_demographie
pw2015_demographie_select = pw2015_demographie %>%
  select(session, short_session, sex, age, sex_orientation, attracted_to, occupation, education,
         sex_orientation, NARQ_scale, NARQ_ADM, NARQ_RIV, BFIK_open, BFIK_consc, BFIK_extra,
         BFIK_agree, BFIK_neuro, created) %>%
  rename(created_t1 = created)

#pw2015_items
pw2015_items_select = pw2015_items %>%
  select(session, short_session, 
          number_relationships, previous_relationship_ended,
          SOI_R_scale, global_interest_LT, global_interest_ST,  longterm_f, longterm_m)

#right censoring: if single, then "created_single_t2" as "start"
t2_t3_relationship <- t2_t3_relationship %>% mutate(
  start = if_else(is.na(start), created_singles_t2, start))

#right censoring: if single, then "created_single_t2" as "start"
t2_t3_relationship_exclusive <- t2_t3_relationship_exclusive %>% mutate(
  start = if_else(is.na(start), created_singles_t2, start))


# use relationship information to merge data on
data = left_join(t2_t3_relationship, pw2015_demographie_select,
                 by = c("session", "short_session"))
data = left_join(data, pw2015_items_select, by = c("session", "short_session"))

data_exclusive = left_join(t2_t3_relationship_exclusive, pw2015_demographie_select,
                 by = c("session", "short_session"))
data_exclusive = left_join(data_exclusive, pw2015_items_select, by = c("session", "short_session"))
```

## Create additional variables for survival analysis (start of relationships)
```{r}
# Create event-variable
data <- data %>% 
  mutate(event = ifelse(lasting == 0, 0, 1),
         created_t1 = as.Date(created_t1),
         # creation of single_start, first we look at people who stayed single during the study
         # their single_start date is the end date of their last relationship:
         single_start = if_else(relationship_order == 0, previous_relationship_ended,
                                previous_relationship_ended),
         # if the information about their previous relationship is missing (was optional in t1),
         # we can simply take the creation date of t1 - it does not make a difference for these
         # people, because they stay single for the whole time. this is a conservative way: we
         # can be sure that they were single at t1, we dont know how long they were single before
         single_start = if_else(relationship_order == 0 & is.na(previous_relationship_ended),
                                created_t1, single_start),
         # the next step are all first relationships: their single_start date should be the end
         # of their previous relationship as well
         single_start = if_else(relationship_order == 1, previous_relationship_ended,
                               single_start),
         # we have the same problem with first relationships: missing previous_relationship_ended
         # data. To deal with this I use the created_t1 as well. This doesnt seem adequat though,
         # because people could have been single for a longer time. Maybe it is conservative, too?
         # we dont know anything else about these people...
         single_start = if_else(relationship_order == 1 & is.na(previous_relationship_ended),
                                created_t1, single_start),
         single_start = if_else(is.na(single_start), created_t1, single_start))

data_exclusive <- data_exclusive %>% 
  mutate(event = ifelse(lasting == 0, 0, 1),
         created_t1 = as.Date(created_t1),
         # creation of single_start, first we look at people who stayed single during the study
         # their single_start date is the end date of their last relationship:
         single_start = if_else(relationship_order == 0, previous_relationship_ended,
                                previous_relationship_ended),
         # if the information about their previous relationship is missing (was optional in t1),
         # we can simply take the creation date of t1 - it does not make a difference for these
         # people, because they stay single for the whole time. this is a conservative way: we
         # can be sure that they were single at t1, we dont know how long they were single before
         single_start = if_else(relationship_order == 0 & is.na(previous_relationship_ended),
                                created_t1, single_start),
         # the next step are all first relationships: their single_start date should be the end
         # of their previous relationship as well
         single_start = if_else(relationship_order == 1, previous_relationship_ended,
                               single_start),
         # we have the same problem with first relationships: missing previous_relationship_ended
         # data. To deal with this I use the created_t1 as well. This doesnt seem adequat though,
         # because people could have been single for a longer time. Maybe it is conservative, too?
         # we dont know anything else about these people...
         single_start = if_else(relationship_order == 1 & is.na(previous_relationship_ended),
                                created_t1, single_start),
         single_start = if_else(is.na(single_start), created_t1, single_start))
# For right now all people have single_start based on their very first relationship or the date of t1 We will fix this in time but for now it should work for the analysis.

# For all other information we have to bring the date back into wideformat
# data_wide = data %>%
#   ungroup() %>%
#   spread()
##### this is where i am trying to find a way right now...


```

##Create additional variables for survival analysis (end of relationships)
```{r}
#event_end variable indicates if relationship has ended. If participants were singles the whole time, NA will be assigned
data <- data %>% 
  mutate(event_end = ifelse(lasting == 2, 1, 
                      ifelse(lasting == 1, 0, NA)))

data_exclusive <- data_exclusive %>% 
  mutate(event_end = ifelse(lasting == 2, 1, 
                      ifelse(lasting == 1, 0, NA)))


################################
#create end variable: either the date of the relationship end or (if right-censored data) the created-date of t2/t3
#!!!!!Why is it called created_singles_t2, wouldn't "t2_created" be more accurate?


data <- data %>% mutate(end = as.Date(end), t3_created = as.Date(t3_created), created_singles_t2 = 
                    as.Date(created_singles_t2),
                  end = if_else(event_end == 0 & is.na(t3_created), created_singles_t2, end),
                  end = if_else(event_end == 0 & !is.na(t3_created), t3_created, end))

data_exclusive <- data_exclusive %>% 
            mutate(end = as.Date(end), t3_created = as.Date(t3_created), created_singles_t2 = 
                    as.Date(created_singles_t2),
                  end = if_else(event_end == 0 & is.na(t3_created), created_singles_t2, end),
                  end = if_else(event_end == 0 & !is.na(t3_created), t3_created, end))

  
```


## Clean data set
```{r}
# clean data set without liars
alt_data_clean = data %>% filter(!(session %in% non_singles_vector_unique))

alt_data_exclusive_clean = data_exclusive %>% filter(!(session %in% non_singles_vector_unique))
```

## Save Data
```{r}
write.csv(data, file = "ImportedData/data.csv")
write.csv(data_clean, file = "ImportedData/data_clean.csv")

write.csv(data_exclusive, file = "ImportedData/data_exclusive.csv")
write.csv(data_exclusive_clean, file = "ImportedData/data_exclusive_clean.csv")

```
