==> 04_Mail2Therneau.Rmd <==
---
title: "04_Mail2Therneau"
author: "Laura Botzet, Sarah Forsthoff, Tanja Gerlach"
date: "28 October 2018"
output: word_document
---
#How to interpret continous predictors in CoxPH?
* We need to interpret them by unit: e.g. one raised unit is (or not) associated with the giving likelihood (see output: exp(coef) which is the hazard ratio, which is hazard over observed time interval)
* How do we scale our data, so we can interpret results neatly? 

#CoxPH refers to the assumption of proportional hazard over different intervals. Does our data fit to this assumption?
check "https://www.mwsug.org/proceedings/2006/stats/MWSUG-2006-SD08.pdf"
*Narcissism's influence declines by age? 
*Test with Schoenfeld residuals --> cox.zph() (see also "https://www.r-bloggers.com/cox-model-assumptions/")
  -> see 03_dataanalyses, I tested this examplatory for one model
*Or graphically with ggcoxzph() in survminer-package
* if violated, try Weighted Cox Regression
("https://www.jstatsoft.org/article/view/v084i02/v84i02.pdf")

#How can we plot results nicely?
*found little about plotting continous predictors (in survival analyses) nicely
*We propably need to bin data, any good ideas? 
*With enough we could use coxPH function on continous and on latent classes, and then plot results for latent classes



#Contacting Mr. Therneau
##Mail address:
therneau@mayo.edu

Dear Mr. Therneau,

We are contacting you as two student assistants working for the department of biological personality psychology at the Georg-August University in Göttingen (Germany). At the moment we are investigating how narcissism is associated with the entry into and the dissolution of romantic relationships. To analyze our longitudinal data, we are using survival analyses or. the coxph-function. 

Because our main predictors (narcissism and its sub-dimensions based on the NARQ-questionnaire) are continous we are struggling with interpreting our results. 

1. Is our interpretation correct based on the given output?

###############################
Call:
coxph(formula = Surv(single_start_d, single_end_d, event_start) ~ 
    NARQ_ADM + NARQ_RIV + sex + age + cluster(session), data = surv_data)

  n= 848, number of events= 305 

              coef exp(coef)  se(coef) robust se      z Pr(>|z|)    
NARQ_ADM  0.347611  1.415681  0.096927  0.104591  3.324 0.000889 ***
NARQ_RIV -0.167863  0.845469  0.098916  0.114159 -1.470 0.141445    
sex      -0.139808  0.869525  0.125437  0.129171 -1.082 0.279095    
age      -0.057734  0.943901  0.008379  0.008411 -6.864 6.71e-12 ***
---
Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

         exp(coef) exp(-coef) lower .95 upper .95
NARQ_ADM    1.4157     0.7064    1.1533    1.7378
NARQ_RIV    0.8455     1.1828    0.6760    1.0575
sex         0.8695     1.1501    0.6750    1.1200
age         0.9439     1.0594    0.9285    0.9596

Concordance= 0.624  (se = 0.018 )
Rsquare= 0.096   (max possible= 0.987 )
Likelihood ratio test= 85.27  on 4 df,   p=<2e-16
Wald test            = 71.19  on 4 df,   p=1e-14
Score (logrank) test = 69.36  on 4 df,   p=3e-14,   Robust = 73.8  p=4e-15

  (Note: the likelihood ratio and score tests assume independence of
     observations within a cluster, the Wald and robust score tests do not).
##################################

"One raised unit of NARQ_ADM is associated with 1.33 higher probability of the event (i.g. starting a relationship)."


2. Would you recommend a different approach for continous data?
Either a different analysis or would you recommend to group our continous data?

3. Can you recommend a good way of plotting our results (because Kaplan-Meier-curves are hard to interpret)?


Best regards,
Laura Botzet & Sarah Forsthoff
